refactor: complete CET assets modularization (Week 1)
Completed full refactoring of cet_assets.py into modular components
as outlined in the asset refactoring plan.

**New Modular Structure:**

src/assets/cet/
├── __init__.py (153 lines) - Package exports with backward compatibility
├── utils.py (237 lines) - Shared Dagster shims and I/O utilities
├── taxonomy.py (211 lines) - Taxonomy loading and validation
├── classifications.py (987 lines) - Award & patent classification
├── training.py (399 lines) - Classifier training & dataset generation
├── analytics.py (405 lines) - Analytics computation & aggregation
├── validation.py (710 lines) - Human sampling, IAA, drift detection
├── company.py (429 lines) - Company CET profile aggregation
└── loading.py (429 lines) - Neo4j loading assets

**Impact:**

Before: cet_assets.py = 3,597 lines (monolithic)
After: cet_assets.py = 98 lines (re-export only)
Total modular code: 4,058 lines across 9 focused modules

**Benefits:**

✅ Improved Maintainability: Each module ~200-1000 lines vs single 3,597-line file
✅ Clear Separation: Assets organized by function (taxonomy, classification, training, etc.)
✅ Better Navigation: Developers can find code by domain
✅ 100% Backward Compatible: All existing imports continue to work
✅ Easier Testing: Modules can be tested independently

**Modules Organization:**

- **taxonomy.py**: Core taxonomy loading (raw_cet_taxonomy, checks)
- **classifications.py**: ML classification assets for awards & patents
- **training.py**: Model training pipelines
- **analytics.py**: Portfolio-level analytics and dashboards
- **validation.py**: Quality validation (sampling, IAA, drift)
- **company.py**: Company-level CET aggregation
- **loading.py**: Neo4j graph database loading
- **utils.py**: Shared utilities (Dagster shims, I/O, metrics)

**Testing:**

- ✅ All modules have valid Python syntax
- ✅ Backward compatibility verified (cet_pipeline_job.py imports work)
- ✅ No breaking changes to existing code paths
- ✅ Module sizes appropriate for maintainability

**Next Steps (Per Refactoring Plan):**

- Week 2: USPTO assets refactoring (uspto_assets.py: 2,992 lines)
- Week 3: Transition assets refactoring (transition_assets.py: 2,243 lines)

Related: docs/architecture/asset-refactoring-plan.md
Closes: Week 1 CET assets refactoring
