+++ src/assets/cet_assets.py
@@ -1,3389 +1,98 @@
 """
 Dagster assets for CET (Critical & Emerging Technologies) taxonomy.
 
-Primary deliverable:
-- `cet_taxonomy` Dagster asset: loads taxonomy via `TaxonomyLoader`,
-  validates it using Pydantic, converts to a DataFrame and writes to
-  `data/processed/cet_taxonomy.parquet` and a companion checks JSON.
+This module re-exports all CET assets from the modularized src.assets.cet package.
+For implementation details, see the individual modules in src/assets/cet/:
+- taxonomy.py: Taxonomy loading and validation
+- classifications.py: Award and patent classification
+- training.py: Classifier training and dataset generation
+- analytics.py: Analytics computation and aggregation
+- validation.py: Human sampling, IAA reports, and drift detection
+- company.py: Company CET profile aggregation
+- loading.py: Neo4j loading assets
+- utils.py: Shared utilities and Dagster shims
 
-Helper functions:
-- `taxonomy_to_dataframe` - convert TaxonomyConfig -> pandas.DataFrame
-- `save_dataframe_parquet` - safe parquet saver that creates directories
-
-Notes:
-- This module intentionally keeps asset logic small and testable so it can
-  be executed in CI without heavy dependencies. File I/O uses pandas'
-  `to_parquet` which requires pyarrow or fastparquet available in the env.
+For backward compatibility, all assets are re-exported at the top level.
 """
 
 from __future__ import annotations
 
-import json
-import os
-from pathlib import Path
-from typing import Any
-
-
-# Import-safe shims for Dagster asset checks
-try:
-    from dagster import (
-        AssetCheckResult,
-        AssetCheckSeverity,
-        AssetIn,
-        asset,
-        asset_check,
-    )  # type: ignore
-    from dagster import AssetExecutionContext as _RealAssetExecutionContext
-
-    # Wrap the real AssetExecutionContext to accept no args for testing
-    class AssetExecutionContext:  # type: ignore
-        def __init__(self, op_execution_context=None, op_config=None) -> None:
-            if op_execution_context is None:
-                # For testing: create a minimal mock-like object with loguru
-                from loguru import logger as _logger
-
-                class _L:
-                    def info(self, *a, **kw):  # noqa: D401
-                        _logger.info(*a, **kw)
-
-                    def warning(self, *a, **kw):
-                        _logger.warning(*a, **kw)
-
-                    def error(self, *a, **kw):
-                        _logger.error(*a, **kw)
-
-                self.log = _L()
-                self.op_config = op_config or {}
-                self._is_shim = True
-            else:
-                # For real usage: use the real Dagster context
-                self._real_context = _RealAssetExecutionContext(op_execution_context)
-                self.log = self._real_context.log
-                # Try to get op_config from the real context if available
-                if hasattr(self._real_context, "op_config"):
-                    self.op_config = self._real_context.op_config
-                elif hasattr(op_execution_context, "op_config"):
-                    self.op_config = op_execution_context.op_config
-                else:
-                    self.op_config = op_config or {}
-                self._is_shim = False
-
-except Exception:  # pragma: no cover
-
-    def asset(*args, **kwargs):  # type: ignore
-        def _wrap(fn):
-            return fn
-
-        return _wrap
-
-    def asset_check(*args, **kwargs):  # type: ignore
-        def _wrap(fn):
-            return fn
-
-        return _wrap
-
-    def AssetIn(*args, **kwargs):  # type: ignore
-        return None
-
-    class AssetCheckResult:  # type: ignore
-        def __init__(self, passed: bool, severity=None, description=None, metadata=None):
-            self.passed = passed
-            self.severity = severity
-            self.description = description
-            self.metadata = metadata
-
-    class AssetCheckSeverity:  # type: ignore
-        ERROR = "ERROR"
-        WARN = "WARN"
-
-    class AssetExecutionContext:  # type: ignore
-        def __init__(self, op_execution_context=None) -> None:
-            from loguru import logger as _logger
-
-            class _L:
-                def info(self, *a, **kw):  # noqa: D401
-                    _logger.info(*a, **kw)
-
-                def warning(self, *a, **kw):
-                    _logger.warning(*a, **kw)
-
-                def error(self, *a, **kw):
-                    _logger.error(*a, **kw)
-
-            self.log = _L()
-            if op_execution_context:
-                # Store if provided for compatibility
-                self._op_execution_context = op_execution_context
-
-
-# Statistical reporting imports
-try:  # pragma: no cover - defensive import
-    from ..models.quality import ModuleReport  # type: ignore
-    from ..utils.reporting.analyzers.cet_analyzer import CetClassificationAnalyzer  # type: ignore
-except Exception:
-    ModuleReport = None  # type: ignore
-    CetClassificationAnalyzer = None  # type: ignore
-
-
-# cet_taxonomy_completeness_check is now imported from .cet.taxonomy module
-# (Original definition removed - see src/assets/cet/taxonomy.py)
-
-
-@asset_check(
-    asset="enriched_cet_award_classifications",
-    description="Award classification quality thresholds (high confidence, evidence coverage) from checks JSON",
-)
-def cet_award_classifications_quality_check(context) -> AssetCheckResult:
-    """
-    Validate CET award classification quality against targets.
-    Consumes data/processed/cet_award_classifications.checks.json written by the asset.
-    """
-    import json
-    import os
-    from pathlib import Path
-
-    checks_path = Path("data/processed/cet_award_classifications.checks.json")
-    if not checks_path.exists():
-        desc = "Missing award classification checks JSON; classification asset may not have run."
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path), "reason": "missing_checks"},
-        )
-
-    try:
-        with checks_path.open("r", encoding="utf-8") as fh:
-            checks = json.load(fh)
-    except Exception as exc:
-        desc = f"Failed to read award classification checks JSON: {exc}"
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path)},
-        )
-
-    # Targets (defaults align with project guidance; override via env)
-    target_high_conf = float(
-        os.environ.get("SBIR_ETL__CET__CLASSIFICATION__HIGH_CONF_THRESHOLD", "0.60")
-    )
-    target_evidence_cov = float(
-        os.environ.get("SBIR_ETL__CET__CLASSIFICATION__EVIDENCE_COVERAGE_THRESHOLD", "0.80")
-    )
-
-    high_conf_rate = checks.get("high_conf_rate")
-    evidence_cov_rate = checks.get("evidence_coverage_rate")
-    model_reason = checks.get("reason")
-
-    # If model missing/failed to load, fail loudly
-    if model_reason in {"model_missing", "model_load_failed"}:
-        desc = f"Classification invalid: {model_reason}"
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path), **checks},
-        )
-
-    # If metrics are missing, warn/fail
-    metrics_present = (high_conf_rate is not None) and (evidence_cov_rate is not None)
-    if not metrics_present:
-        desc = "Classification checks JSON missing quality metrics"
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path), **checks},
-        )
-
-    passed = (high_conf_rate >= target_high_conf) and (evidence_cov_rate >= target_evidence_cov)
-    desc = (
-        "Classification quality meets thresholds"
-        if passed
-        else "Classification quality below thresholds"
-    )
-    severity = AssetCheckSeverity.WARN if passed else AssetCheckSeverity.ERROR
-    metadata = {
-        "checks_path": str(checks_path),
-        "high_conf_rate": high_conf_rate,
-        "evidence_coverage_rate": evidence_cov_rate,
-        "target_high_conf_rate": target_high_conf,
-        "target_evidence_coverage_rate": target_evidence_cov,
-        **{
-            k: v for k, v in checks.items() if k not in {"high_conf_rate", "evidence_coverage_rate"}
-        },
-    }
-    return AssetCheckResult(passed=passed, severity=severity, description=desc, metadata=metadata)
-
-
-@asset_check(
-    asset="transformed_cet_company_profiles",
-    description="Company CET profiles successfully generated (basic sanity from checks JSON)",
-)
-def cet_company_profiles_check(context) -> AssetCheckResult:
-    """
-    Ensure company CET profiles were produced without critical errors.
-    Consumes data/processed/cet_company_profiles.checks.json written by the asset.
-    """
-    import json
-    from pathlib import Path
-
-    checks_path = Path("data/processed/cet_company_profiles.checks.json")
-    if not checks_path.exists():
-        desc = "Missing company profiles checks JSON; aggregation asset may not have run."
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path), "reason": "missing_checks"},
-        )
-
-    try:
-        with checks_path.open("r", encoding="utf-8") as fh:
-            checks = json.load(fh)
-    except Exception as exc:
-        desc = f"Failed to read company profiles checks JSON: {exc}"
-        context.log.error(desc)
-        return AssetCheckResult(
-            passed=False,
-            severity=AssetCheckSeverity.ERROR,
-            description=desc,
-            metadata={"checks_path": str(checks_path)},
-        )
-
-    ok = bool(checks.get("ok", False))
-    desc = "Company profile generation passed" if ok else "Company profile generation failed"
-    severity = AssetCheckSeverity.WARN if ok else AssetCheckSeverity.ERROR
-    return AssetCheckResult(
-        passed=ok,
-        severity=severity,
-        description=desc,
-        metadata={"checks_path": str(checks_path), **checks},
-    )
-
-
-from collections.abc import Iterable
-
-import pandas as pd
-
-
-try:
-    from dagster import AssetKey, MetadataValue, Output, asset
-except Exception:  # pragma: no cover - fallback stubs when dagster is not installed
-    # Lightweight stubs so this module can be imported in environments without dagster.
-    class Output:
-        def __init__(self, value, metadata=None):
-            self.value = value
-            self.metadata = metadata
-
-        def __repr__(self):
-            return f"Output(value={self.value!r}, metadata={self.metadata!r})"
-
-    def asset(*dargs, **dkwargs):
-        # decorator passthrough: return the function unchanged
-        def _decorator(fn):
-            return fn
-
-        return _decorator
-
-    class AssetKey:
-        def __init__(self, key):
-            self.key = key
-
-        def __repr__(self):
-            return f"AssetKey({self.key!r})"
-
-    class MetadataValue:
-        @staticmethod
-        def text(s):
-            return s
-
-
-from loguru import logger
-
-from src.ml.config.taxonomy_loader import TaxonomyLoader
-from src.models.cet_models import CETArea
-
-# Import from modularized CET subpackage
+# Import all assets from the modularized cet package
 from .cet import (
+    # Taxonomy
     cet_taxonomy,
     cet_taxonomy_completeness_check,
     raw_cet_taxonomy,
-    save_dataframe_parquet,
     taxonomy_to_dataframe,
+    # Classifications
+    cet_award_classifications_quality_check,
+    enriched_cet_award_classifications,
+    enriched_cet_patent_classifications,
+    # Training
+    cet_award_training_dataset,
+    train_cet_patent_classifier,
+    # Analytics
+    transformed_cet_analytics,
+    transformed_cet_analytics_aggregates,
+    # Validation
+    raw_cet_human_sampling,
+    validated_cet_drift_detection,
+    validated_cet_iaa_report,
+    # Company
+    cet_company_profiles_check,
+    transformed_cet_company_profiles,
+    # Loading
+    loaded_award_cet_enrichment,
+    loaded_award_cet_relationships,
+    loaded_cet_areas,
+    loaded_company_cet_enrichment,
+    loaded_company_cet_relationships,
+    # Backward compatibility aliases
+    neo4j_award_cet_enrichment,
+    neo4j_award_cet_relationships,
+    neo4j_cetarea_nodes,
+    neo4j_company_cet_enrichment,
+    neo4j_company_cet_relationships,
+    # Utility functions
+    save_dataframe_parquet,
 )
 
 
-# Default output location for processed taxonomy
-DEFAULT_OUTPUT_PATH = Path("data/processed/cet_taxonomy.parquet")
-
-
-# taxonomy_to_dataframe and save_dataframe_parquet are now imported from .cet module
-
-
-# raw_cet_taxonomy and cet_taxonomy are now imported from .cet.taxonomy module
-# (Original definitions removed - see src/assets/cet/taxonomy.py)
-
-
-@asset(
-    name="enriched_cet_award_classifications",
-    key_prefix=["ml"],
-    description=(
-        "Batch classify SBIR awards with the trained CET classifier, extract evidence "
-        "per classification, persist results to `data/processed/cet_award_classifications.parquet` "
-        "and emit a companion checks JSON for automated validation."
-    ),
-)
-def enriched_cet_award_classifications() -> Output:
-    """
-    Dagster asset to perform batch CET classification over enriched award records.
-
-    Behavior (best-effort / import-safe):
-    - Attempts to load a trained ApplicabilityModel from a well-known artifacts path.
-      If the model is missing, writes an empty classification output and a checks JSON
-      explaining the missing artifact.
-    - Loads the taxonomy via TaxonomyLoader and instantiates EvidenceExtractor.
-    - Reads enriched award records from `data/processed/enriched_sbir_awards.ndjson`
-      (NDJSON) or `data/processed/enriched_sbir_awards.parquet` if available. If neither
-      exists, operates on a very small sample to allow the asset to run in CI.
-    - Classifies awards in batches (configurable via classification config) and attaches
-      up to N evidence statements per CET classification.
-    - Persists classifications to `data/processed/cet_award_classifications.parquet`
-      using the same parquet -> NDJSON fallback approach as the taxonomy asset.
-    - Writes a checks JSON summarizing classification coverage and confidence rates.
-    """
-    logger.info("Starting cet_award_classifications asset")
-
-    # Local imports to keep module import-safe when optional deps are missing
-    import json
-    from pathlib import Path
-
-    # Lazy imports for ML components (may be unavailable in minimal CI)
-    try:
-        from src.ml.features.evidence_extractor import EvidenceExtractor
-    except Exception:
-        EvidenceExtractor = None  # type: ignore
-
-    try:
-        from src.ml.models.cet_classifier import ApplicabilityModel
-    except Exception:
-        ApplicabilityModel = None  # type: ignore
-
-    # Paths and defaults
-    awards_ndjson = Path("data/processed/enriched_sbir_awards.ndjson")
-    awards_parquet = Path("data/processed/enriched_sbir_awards.parquet")
-    model_path = Path("artifacts/models/cet_classifier_v1.pkl")
-    output_path = Path("data/processed/cet_award_classifications.parquet")
-    checks_path = output_path.with_suffix(".checks.json")
-
-    # Load taxonomy and classification config (required for EvidenceExtractor and thresholds)
-    try:
-        loader = TaxonomyLoader()
-        taxonomy = loader.load_taxonomy()
-    except Exception:
-        logger.exception("Failed to load taxonomy; writing empty output")
-        import pandas as pd
-
-        df_empty = pd.DataFrame(
-            columns=[
-                "award_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "evidence",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        try:
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            out_json = output_path.with_suffix(".json")
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-        checks = {"ok": False, "reason": "taxonomy_load_failed", "num_awards": 0}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    try:
-        classification_config = loader.load_classification_config()
-    except Exception:
-        # If classification config cannot be loaded, fall back to defaults
-        logger.warning("Could not load classification config; using defaults")
-        classification_config = {}
-
-    # Prepare EvidenceExtractor if available
-    extractor = None
-    if EvidenceExtractor is not None:
-        try:
-            extractor = EvidenceExtractor(list(taxonomy.cet_areas), classification_config)
-        except Exception:
-            extractor = None
-            logger.exception("Failed to initialize EvidenceExtractor; evidence extraction disabled")
-
-    # Load awards (prefer parquet, then ndjson). If neither present, use a tiny sample.
-    awards: list[dict] = []
-    try:
-        if awards_parquet.exists():
-            import pandas as pd
-
-            df_awards = pd.read_parquet(awards_parquet)
-            # Expect dataframe with at least award_id/title/abstract/keywords
-            for _, row in df_awards.iterrows():
-                awards.append(
-                    {
-                        "award_id": str(row.get("award_id") or row.get("id") or ""),
-                        "title": str(row.get("title") or ""),
-                        "abstract": str(row.get("abstract") or ""),
-                        "keywords": row.get("keywords") or "",
-                    }
-                )
-        elif awards_ndjson.exists():
-            with open(awards_ndjson, encoding="utf-8") as fh:
-                for line in fh:
-                    if line.strip():
-                        awards.append(json.loads(line))
-        else:
-            # Minimal sample so asset can run in lightweight CI
-            awards = [
-                {
-                    "award_id": "sample_001",
-                    "title": "AI for imaging",
-                    "abstract": "This project applies machine learning and deep neural networks to image analysis.",
-                    "keywords": ["machine learning", "neural networks"],
-                },
-                {
-                    "award_id": "sample_002",
-                    "title": "Quantum algorithms",
-                    "abstract": "Research on quantum optimization and qubit coherence for algorithms.",
-                    "keywords": ["quantum computing", "qubits"],
-                },
-            ]
-            logger.warning(
-                "No enriched awards found at expected paths; running classification on a small sample"
-            )
-    except Exception:
-        logger.exception("Failed to load awards for classification; writing empty output")
-        awards = []
-
-    # If model artifact not found or loading fails, write placeholder output & checks
-    if not model_path.exists() or ApplicabilityModel is None:
-        logger.warning(
-            "Trained model not available; skipping classification (model: %s)", model_path
-        )
-        # Produce an empty DataFrame with expected columns so downstream consumers have schema
-        import pandas as pd
-
-        df_empty = pd.DataFrame(
-            columns=[
-                "award_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "evidence",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        # Use existing save helper for parquet/NDJSON fallback
-        try:
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            # If save failed, attempt NDJSON write
-            out_json = output_path.with_suffix(".json")
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-
-        checks = {
-            "ok": False,
-            "reason": "model_missing",
-            "model_path": str(model_path),
-            "num_awards": len(awards),
-            "num_classified": 0,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": False,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Load trained model
-    try:
-        model = ApplicabilityModel.load(model_path)
-    except Exception:
-        logger.exception("Failed to load trained model from %s", model_path)
-        model = None
-
-    if model is None:
-        logger.warning("Model could not be loaded; aborting classification")
-        df_empty = __import__("pandas").DataFrame(
-            columns=[
-                "award_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "evidence",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        save_dataframe_parquet(df_empty, output_path)
-        checks = {"ok": False, "reason": "model_load_failed", "num_awards": len(awards)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": False,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Build texts for classification and perform batch classification
-    texts = []
-    award_ids = []
-    for a in awards:
-        combined = " ".join(
-            filter(
-                None,
-                [
-                    str(a.get("title", "")),
-                    str(a.get("abstract", "")),
-                    " ".join(a.get("keywords") or []),
-                ],
-            )
-        )
-        texts.append(combined)
-        award_ids.append(a.get("award_id") or "")
-
-    # Handle empty awards list
-    if not texts or not award_ids:
-        logger.warning("No award texts to classify; writing empty output")
-        df_empty = pd.DataFrame(
-            columns=[
-                "award_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "evidence",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        try:
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            out_json = output_path.with_suffix(".json")
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-        checks = {"ok": False, "reason": "no_awards_to_classify", "num_awards": len(awards)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": True,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    batch_size = (
-        classification_config.get("batch", {}).get("size", 1000)
-        if isinstance(classification_config, dict)
-        else 1000
-    )
-
-    # Perform batch classification with error handling
-    try:
-        classifications_by_award = model.classify_batch(texts, batch_size=batch_size)
-    except Exception:
-        logger.exception("Batch classification failed; writing empty output")
-        df_empty = pd.DataFrame(
-            columns=[
-                "award_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "evidence",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        try:
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            out_json = output_path.with_suffix(".json")
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-        checks = {"ok": False, "reason": "classification_failed", "num_awards": len(awards)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": True,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Attach evidence if extractor available
-    if extractor is not None:
-        # Build document_parts list matching expectations of EvidenceExtractor
-        doc_parts_list = []
-        for a in awards:
-            doc_parts_list.append(
-                {
-                    "abstract": str(a.get("abstract", "")),
-                    "keywords": " ".join(a.get("keywords") or []),
-                    "title": str(a.get("title", "")),
-                }
-            )
-        try:
-            classifications_with_evidence = extractor.extract_batch_evidence(
-                classifications_by_award, doc_parts_list
-            )
-        except Exception:
-            logger.exception(
-                "Batch evidence extraction failed; falling back to classification-only results"
-            )
-            classifications_with_evidence = classifications_by_award
-    else:
-        classifications_with_evidence = classifications_by_award
-
-    # Flatten classification results into a DataFrame (one row per award)
-    import pandas as pd
-
-    rows = []
-    for aid, cls_list in zip(award_ids, classifications_with_evidence, strict=False):
-        if not cls_list:
-            rows.append(
-                {
-                    "award_id": aid,
-                    "primary_cet": None,
-                    "primary_score": None,
-                    "supporting_cets": [],
-                    "evidence": [],
-                    "classified_at": None,
-                    "taxonomy_version": model.taxonomy_version if model else taxonomy.version,
-                }
-            )
-            continue
-
-        primary = cls_list[0]
-        supporting = cls_list[1:4] if len(cls_list) > 1 else []
-        rows.append(
-            {
-                "award_id": aid,
-                "primary_cet": primary.cet_id,
-                "primary_score": primary.score,
-                "supporting_cets": [
-                    {
-                        "cet_id": s.cet_id,
-                        "score": s.score,
-                        "classification": s.classification.value
-                        if hasattr(s.classification, "value")
-                        else str(s.classification),
-                    }
-                    for s in supporting
-                ],
-                "evidence": [
-                    {
-                        "excerpt": e.excerpt,
-                        "source": e.source_location,
-                        "rationale": e.rationale_tag,
-                    }
-                    for e in primary.evidence
-                ]
-                if getattr(primary, "evidence", None)
-                else [],
-                "classified_at": primary.classified_at,
-                "taxonomy_version": primary.taxonomy_version,
-            }
-        )
-
-    df_out = pd.DataFrame(rows)
-
-    # Persist classifications (parquet preferred, NDJSON fallback)
-    try:
-        save_dataframe_parquet(df_out, output_path)
-    except Exception:
-        # Fallback: write NDJSON manually
-        json_out = output_path.with_suffix(".json")
-        json_out.parent.mkdir(parents=True, exist_ok=True)
-        with open(json_out, "w", encoding="utf-8") as fh:
-            for rec in rows:
-                fh.write(json.dumps(rec) + "\n")
-        # Touch parquet placeholder for consumers that assert its existence
-        try:
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.touch()
-        except Exception:
-            logger.exception("Failed to touch parquet placeholder file for classifications")
-
-    # Build checks: coverage, high-confidence rate, evidence coverage
-    num_awards = len(rows)
-    num_classified = sum(1 for r in rows if r.get("primary_cet"))
-    # Precompute high confidence threshold to avoid complex inline conditional expressions
-    if isinstance(classification_config, dict):
-        high_threshold = classification_config.get("confidence_thresholds", {}).get("high", 70.0)
-    else:
-        # classification_config may be a Pydantic model; attempt attribute access, fall back to default
-        try:
-            high_threshold = getattr(classification_config, "confidence_thresholds", {}).get(
-                "high", 70.0
-            )
-        except Exception:
-            high_threshold = 70.0
-    high_conf_count = sum(1 for r in rows if (r.get("primary_score") or 0) >= high_threshold)
-    evidence_coverage = sum(1 for r in rows if r.get("evidence"))
-
-    checks = {
-        "ok": True,
-        "num_awards": num_awards,
-        "num_classified": num_classified,
-        "high_conf_count": int(high_conf_count),
-        "high_conf_rate": float(high_conf_count / max(1, num_awards)),
-        "evidence_coverage_count": int(evidence_coverage),
-        "evidence_coverage_rate": float(evidence_coverage / max(1, num_awards)),
-        "model_path": str(model_path),
-    }
-
-    checks_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(checks_path, "w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    metadata = {
-        "path": str(output_path),
-        "rows": len(df_out),
-        "taxonomy_version": taxonomy.version,
-        "model_version": getattr(model, "model_version", None),
-        "checks_path": str(checks_path),
-    }
-
-    logger.info(
-        "Completed cet_award_classifications asset", rows=len(df_out), output=str(output_path)
-    )
-
-    # Perform statistical analysis with CET analyzer
-    if CetClassificationAnalyzer is not None:
-        try:
-            analyzer = CetClassificationAnalyzer()
-            run_context = {
-                "run_id": "cet_classification_run",  # Could be enhanced with actual run_id
-                "pipeline_name": "cet_classification",
-                "stage": "transform",
-            }
-
-            # Load the classified DataFrame from the output file
-            import pandas as pd
-
-            try:
-                classified_df = pd.read_parquet(output_path)
-            except Exception:
-                logger.warning(f"Could not load classified DataFrame from {output_path}")
-                classified_df = pd.DataFrame()
-
-            # Prepare module data for analysis
-            module_data = {
-                "classified_df": classified_df,
-                "classification_results": {
-                    "classified_records": len(classified_df),
-                    "failed_records": 0,  # Assume all records were processed
-                    "duration_seconds": 0.0,  # Could be enhanced with actual timing
-                    "classification_rate": len(
-                        classified_df[classified_df["primary_cet_area"].notna()]
-                    )
-                    / len(classified_df)
-                    if len(classified_df) > 0
-                    else 0,
-                    "model_accuracy": checks.get("accuracy", 0.0),
-                    "model_precision": checks.get("precision", 0.0),
-                    "model_recall": checks.get("recall", 0.0),
-                    "model_f1_score": checks.get("f1_score", 0.0),
-                },
-                "taxonomy_data": {
-                    "taxonomy_areas": list(taxonomy.cet_areas.keys())
-                    if hasattr(taxonomy, "cet_areas")
-                    else [],
-                    "total_areas": len(taxonomy.cet_areas) if hasattr(taxonomy, "cet_areas") else 0,
-                },
-                "run_context": run_context,
-            }
-
-            # Generate analysis report
-            analysis_report = analyzer.analyze(module_data)
-
-            logger.info(
-                "CET classification analysis complete",
-                extra={
-                    "insights_generated": len(analysis_report.insights)
-                    if hasattr(analysis_report, "insights")
-                    else 0,
-                    "data_hygiene_score": analysis_report.data_hygiene.quality_score_mean
-                    if analysis_report.data_hygiene
-                    else None,
-                    "classification_success_rate": analysis_report.success_rate,
-                },
-            )
-
-            # Add analysis results to metadata
-            metadata.update(
-                {
-                    "analysis_insights_count": len(analysis_report.insights)
-                    if hasattr(analysis_report, "insights")
-                    else 0,
-                    "analysis_data_hygiene_score": round(
-                        analysis_report.data_hygiene.quality_score_mean, 3
-                    )
-                    if analysis_report.data_hygiene
-                    else None,
-                    "analysis_category_distribution": analysis_report.module_metrics.get(
-                        "category_distribution", {}
-                    )
-                    if analysis_report.module_metrics
-                    else {},
-                    "analysis_confidence_distribution": analysis_report.module_metrics.get(
-                        "confidence_distribution", {}
-                    )
-                    if analysis_report.module_metrics
-                    else {},
-                    "analysis_taxonomy_coverage": analysis_report.module_metrics.get(
-                        "taxonomy_coverage", {}
-                    )
-                    if analysis_report.module_metrics
-                    else {},
-                }
-            )
-
-        except Exception as e:
-            logger.warning(f"CET classification analysis failed: {e}")
-    else:
-        logger.info("CET analyzer not available; skipping statistical analysis")
-
-    return Output(value=str(output_path), metadata=metadata)
-
-
-@asset(
-    name="enriched_cet_patent_classifications",
-    key_prefix=["ml"],
-    description=(
-        "Batch classify patents with a trained Patent CET classifier, persist results to "
-        "`data/processed/cet_patent_classifications.parquet` and emit a companion checks JSON."
-    ),
-)
-def enriched_cet_patent_classifications() -> Output:
-    """
-    Dagster asset to perform batch CET classification over patent title records.
-
-    Behavior (best-effort / import-safe):
-    - Attempts to load a trained PatentCETClassifier from a well-known artifacts path.
-      If the model is missing, writes an empty classification output and a checks JSON
-      explaining the missing artifact.
-    - Loads the taxonomy via TaxonomyLoader for metadata.
-    - Reads transformed patent records from `data/processed/transformed_patents.ndjson`
-      (NDJSON) or `data/processed/transformed_patents.parquet` if available. If neither
-      exists, operates on a very small sample so the asset can run in CI.
-    - Classifies patent titles in batches and persists outputs with NDJSON/parquet fallback.
-    - Writes a companion checks JSON summarizing classification coverage.
-    """
-    logger.info("Starting cet_patent_classifications asset")
-
-    # Local imports to keep module import-safe when optional deps are missing
-    import json
-    from pathlib import Path
-
-    # Lazy import of classifier implementation (may be unavailable in minimal CI)
-    try:
-        from src.ml.models.patent_classifier import PatentCETClassifier
-    except Exception:
-        PatentCETClassifier = None  # type: ignore
-
-    # Paths and defaults
-    patents_ndjson = Path("data/processed/transformed_patents.ndjson")
-    patents_parquet = Path("data/processed/transformed_patents.parquet")
-    model_path = Path("artifacts/models/patent_classifier_v1.pkl")
-    output_path = Path("data/processed/cet_patent_classifications.parquet")
-    checks_path = output_path.with_suffix(".checks.json")
-
-    # Load taxonomy for metadata (non-fatal)
-    loader = TaxonomyLoader()
-    try:
-        taxonomy = loader.load_taxonomy()
-    except Exception:
-        taxonomy = None
-
-    # Load patent records (prefer parquet, then ndjson). If neither present, use a tiny sample.
-    patents: list[dict] = []
-    try:
-        if patents_parquet.exists():
-            import pandas as pd
-
-            df_patents = pd.read_parquet(patents_parquet)
-            # Expect dataframe with at least patent_id/title/assignee
-            for _, row in df_patents.iterrows():
-                patents.append(
-                    {
-                        "patent_id": str(row.get("patent_id") or row.get("id") or ""),
-                        "title": str(row.get("title") or ""),
-                        "assignee": row.get("assignee") or None,
-                    }
-                )
-        elif patents_ndjson.exists():
-            with open(patents_ndjson, encoding="utf-8") as fh:
-                for line in fh:
-                    if line.strip():
-                        patents.append(json.loads(line))
-        else:
-            # Minimal sample so asset can run in lightweight CI
-            patents = [
-                {
-                    "patent_id": "sample_p1",
-                    "title": "Machine learning for image analysis",
-                    "assignee": "Acme Labs",
-                },
-                {
-                    "patent_id": "sample_p2",
-                    "title": "Quantum computing improvements for qubit stability",
-                    "assignee": "QuantumCorp",
-                },
-            ]
-            logger.warning(
-                "No transformed patents found at expected paths; running classification on a small sample"
-            )
-    except Exception:
-        logger.exception("Failed to load patent records for classification; writing empty output")
-        patents = []
-
-    # If model artifact not found or loading fails, write placeholder output & checks
-    if not model_path.exists() or PatentCETClassifier is None:
-        logger.warning(
-            "Trained patent model not available; skipping classification (model: %s)", model_path
-        )
-        # Produce an empty DataFrame with expected columns so downstream consumers have schema
-        import pandas as pd
-
-        df_empty = pd.DataFrame(
-            columns=[
-                "patent_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        # Use existing save helper for parquet/NDJSON fallback
-        try:
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            # If save failed, attempt NDJSON write
-            out_json = output_path.with_suffix(".json")
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-
-        checks = {
-            "ok": False,
-            "reason": "model_missing",
-            "model_path": str(model_path),
-            "num_patents": len(patents),
-            "num_classified": 0,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": False,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Load trained model
-    try:
-        classifier = PatentCETClassifier.load(model_path)
-    except Exception:
-        logger.exception("Failed to load patent classifier from %s", model_path)
-        classifier = None
-
-    if classifier is None:
-        logger.warning("Patent model could not be loaded; aborting classification")
-        df_empty = __import__("pandas").DataFrame(
-            columns=[
-                "patent_id",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "classified_at",
-                "taxonomy_version",
-            ]
-        )
-        save_dataframe_parquet(df_empty, output_path)
-        checks = {"ok": False, "reason": "model_load_failed", "num_patents": len(patents)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "model_present": False,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Build texts for classification and perform batch classification
-    titles = []
-    patent_ids = []
-    assignees = []
-    # Prefer PatentFeatureExtractor for normalized title strings if available
-    try:
-        from src.ml.features.patent_features import get_keywords_map  # type: ignore
-        from src.ml.models.patent_classifier import PatentFeatureExtractor  # type: ignore
-
-        kw_map = get_keywords_map()
-        extractor = PatentFeatureExtractor(keywords_map=kw_map)
-        if hasattr(extractor, "transform"):
-            feature_dicts = extractor.transform(patents)  # type: ignore
-            for p, fv in zip(patents, feature_dicts, strict=False):
-                norm_title = (
-                    fv.get("normalized_title")
-                    if isinstance(fv, dict)
-                    else getattr(fv, "normalized_title", p.get("title"))
-                )
-                titles.append(str(norm_title or ""))
-                patent_ids.append(p.get("patent_id") or "")
-                # Avoid double-adding assignee when text already normalized
-                assignees.append(None)
-        else:
-            # Fallback to simple normalization when only DF-based extractor is available
-            from src.ml.features.patent_features import normalize_title  # type: ignore
-
-            for p in patents:
-                titles.append(normalize_title(p.get("title")))
-                patent_ids.append(p.get("patent_id") or "")
-                assignees.append(None)
-    except Exception:
-        for p in patents:
-            titles.append(str(p.get("title") or ""))
-            patent_ids.append(p.get("patent_id") or "")
-            assignees.append(p.get("assignee") or None)
-
-    # batch_size: try to read from classification config if loader provided it, else default 1000
-    try:
-        classification_config = loader.load_classification_config()
-    except Exception:
-        classification_config = {}
-
-    batch_size = (
-        classification_config.get("batch", {}).get("size", 1000)
-        if isinstance(classification_config, dict)
-        else 1000
-    )
-
-    classifications_by_patent = classifier.classify_batch(titles, assignees, batch_size=batch_size)
-
-    # Flatten classification results into a DataFrame (one row per patent)
-    import pandas as pd
-
-    rows = []
-    for pid, cls_list in zip(patent_ids, classifications_by_patent, strict=False):
-        if not cls_list:
-            rows.append(
-                {
-                    "patent_id": pid,
-                    "primary_cet": None,
-                    "primary_score": None,
-                    "supporting_cets": [],
-                    "classified_at": None,
-                    "taxonomy_version": classifier.taxonomy_version
-                    if classifier
-                    else (taxonomy.version if taxonomy else None),
-                }
-            )
-            continue
-
-        primary = cls_list[0]
-        supporting = cls_list[1:4] if len(cls_list) > 1 else []
-        rows.append(
-            {
-                "patent_id": pid,
-                "primary_cet": primary.cet_id,
-                "primary_score": primary.score,
-                "supporting_cets": [{"cet_id": s.cet_id, "score": s.score} for s in supporting],
-                "classified_at": getattr(primary, "classified_at", None),
-                "taxonomy_version": getattr(
-                    primary, "taxonomy_version", classifier.taxonomy_version if classifier else None
-                ),
-            }
-        )
-
-    df_out = pd.DataFrame(rows)
-
-    # Persist classifications (parquet preferred, NDJSON fallback)
-    try:
-        save_dataframe_parquet(df_out, output_path)
-    except Exception:
-        # Fallback: write NDJSON manually
-        json_out = output_path.with_suffix(".json")
-        json_out.parent.mkdir(parents=True, exist_ok=True)
-        with open(json_out, "w", encoding="utf-8") as fh:
-            for rec in rows:
-                fh.write(json.dumps(rec) + "\n")
-        # Touch parquet placeholder for consumers that assert its existence
-        try:
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.touch()
-        except Exception:
-            logger.exception("Failed to touch parquet placeholder file for patent classifications")
-
-    # Build checks: coverage and counts
-    num_patents = len(rows)
-    num_classified = sum(1 for r in rows if r.get("primary_cet"))
-    checks = {
-        "ok": True,
-        "num_patents": num_patents,
-        "num_classified": num_classified,
-        "model_path": str(model_path),
-    }
-
-    checks_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(checks_path, "w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    metadata = {
-        "path": str(output_path),
-        "rows": len(df_out),
-        "taxonomy_version": taxonomy.version if taxonomy else None,
-        "model_version": getattr(classifier, "model_version", None),
-        "checks_path": str(checks_path),
-    }
-
-    logger.info(
-        "Completed cet_patent_classifications asset", rows=len(df_out), output=str(output_path)
-    )
-
-    return Output(value=str(output_path), metadata=metadata)
-
-
-@asset(
-    name="train_cet_patent_classifier",
-    key_prefix=["ml"],
-    description=(
-        "Train and persist a Patent CET classifier artifact at "
-        "`artifacts/models/patent_classifier_v1.pkl`. Emits a companion checks JSON "
-        "with training metadata. If training data or dependencies are missing, "
-        "writes a minimal checks file and returns the intended model path."
-    ),
-)
-def train_cet_patent_classifier() -> Output:
-    """
-    Dagster asset that trains the PatentCETClassifier from a labeled dataset.
-
-    Behavior (import-safe):
-    - Expects a labeled training dataset at `data/processed/cet_patent_training.parquet`
-      with columns: `title`, `cet_labels` (iterable[str] or comma-delimited), and optional `assignee`.
-    - When training data or pandas is missing, writes a checks JSON with ok=False and
-      returns the model path without creating the artifact.
-    """
-    import json
-    from pathlib import Path
-
-    model_path = Path("artifacts/models/patent_classifier_v1.pkl")
-    checks_path = model_path.with_suffix(".checks.json")
-    train_data_parquet = Path("data/processed/cet_patent_training.parquet")
-
-    try:
-        import pandas as pd  # type: ignore
-    except Exception:
-        # Missing pandas; write checks and exit
-        checks = {"ok": False, "reason": "pandas_missing", "model_path": str(model_path)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        return Output(
-            value=str(model_path), metadata={"model_path": str(model_path), "trained": False}
-        )
-
-    # Attempt to import training helper and dummy pipeline for simple factory
-    try:
-        from src.ml.features.patent_features import get_keywords_map
-        from src.ml.models.dummy_pipeline import DummyPipeline
-        from src.ml.train.patent_training import train_patent_classifier
-    except Exception:
-        checks = {"ok": False, "reason": "training_helper_missing", "model_path": str(model_path)}
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        return Output(
-            value=str(model_path), metadata={"model_path": str(model_path), "trained": False}
-        )
-
-    if not train_data_parquet.exists():
-        checks = {
-            "ok": False,
-            "reason": "training_data_missing",
-            "expected_path": str(train_data_parquet),
-            "model_path": str(model_path),
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        return Output(
-            value=str(model_path), metadata={"model_path": str(model_path), "trained": False}
-        )
-
-    # Load training data
-    try:
-        df = pd.read_parquet(train_data_parquet)
-    except Exception:
-        # Try NDJSON fallback
-        ndjson_path = train_data_parquet.with_suffix(".ndjson")
-        records = []
-        if ndjson_path.exists():
-            with open(ndjson_path, encoding="utf-8") as fh:
-                for line in fh:
-                    if line.strip():
-                        records.append(json.loads(line))
-        df = pd.DataFrame(records)
-
-    if df is None or len(df) == 0:
-        checks = {
-            "ok": False,
-            "reason": "empty_training_data",
-            "model_path": str(model_path),
-            "rows": 0,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        return Output(
-            value=str(model_path), metadata={"model_path": str(model_path), "trained": False}
-        )
-
-    # Provide a simple pipelines factory using DummyPipeline with CET id as keyword cue
-    def _factory(cet_id: str):
-        # Heuristic: derive a token from CET id for keyword; this keeps CI deterministic
-        kw = cet_id.replace("_", " ")
-        return DummyPipeline(cet_id=cet_id, keywords=[kw], keyword_boost=1.0)
-
-    # Train and persist
-    try:
-        meta = train_patent_classifier(
-            df=df,
-            output_model_path=model_path,
-            pipelines_factory=_factory,
-            title_col="title",
-            assignee_col="assignee" if "assignee" in df.columns else None,
-            cet_label_col="cet_labels",
-            use_feature_extraction=True,
-            keywords_map=get_keywords_map(),
-        )
-        checks = {
-            "ok": True,
-            "model_path": str(model_path),
-            "trained_on_rows": meta.get("trained_on_rows", 0),
-        }
-    except Exception as e:
-        checks = {
-            "ok": False,
-            "reason": "training_failed",
-            "error": str(e),
-            "model_path": str(model_path),
-        }
-
-    checks_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(checks_path, "w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    metadata = {
-        "model_path": str(model_path),
-        "checks_path": str(checks_path),
-        "trained": checks.get("ok", False),
-    }
-    return Output(value=str(model_path), metadata=metadata)
-
-
-# New asset: cet_award_training_dataset
-@asset(
-    name="cet_award_training_dataset",
-    key_prefix=["ml"],
-    description=(
-        "Load labeled CET award training dataset from CSV or NDJSON, validate and persist to "
-        "`data/processed/cet_award_training.parquet` and emit a companion checks JSON. "
-        "Import-safe with NDJSON fallback when parquet engine or pandas is unavailable."
-    ),
-)
-def cet_award_training_dataset() -> Output:
-    import json
-    from pathlib import Path
-
-    output_path = Path("data/processed/cet_award_training.parquet")
-    checks_path = output_path.with_suffix(".checks.json")
-
-    # Candidate input paths (prefer processed)
-    candidate_inputs = [
-        Path("data/processed/cet_award_training.ndjson"),
-        Path("data/processed/cet_award_training.jsonl"),
-        Path("data/processed/cet_award_training.csv"),
-        Path("data/raw/cet_award_training.ndjson"),
-        Path("data/raw/cet_award_training.jsonl"),
-        Path("data/raw/cet_award_training.csv"),
-    ]
-    input_path = next((p for p in candidate_inputs if p.exists()), None)
-
-    # Load taxonomy for metadata
-    loader = TaxonomyLoader()
-    try:
-        taxonomy = loader.load_taxonomy()
-        taxonomy_version = taxonomy.version
-    except Exception:
-        taxonomy = None
-        taxonomy_version = None
-
-    # If no input, write empty output and checks
-    if input_path is None:
-        try:
-            import pandas as pd  # type: ignore
-
-            df_empty = pd.DataFrame(
-                columns=[
-                    "example_id",
-                    "text",
-                    "title",
-                    "keywords",
-                    "keywords_joined",
-                    "solicitation",
-                    "labels",
-                    "labels_joined",
-                    "source",
-                    "annotated_by",
-                    "annotated_at",
-                    "notes",
-                    "taxonomy_version",
-                ]
-            )
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            # Ensure a placeholder parquet file exists and write empty NDJSON
-            out_json = output_path.with_suffix(".ndjson")
-            out_json.parent.mkdir(parents=True, exist_ok=True)
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-            try:
-                output_path.parent.mkdir(parents=True, exist_ok=True)
-                output_path.touch()
-            except Exception:
-                pass
-
-        checks = {
-            "ok": False,
-            "reason": "training_data_missing",
-            "expected_paths": [str(p) for p in candidate_inputs],
-            "rows": 0,
-            "taxonomy_version": taxonomy_version,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "checks_path": str(checks_path),
-            "input_path": None,
-            "taxonomy_version": taxonomy_version,
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Load dataset using the training loader
-    try:
-        from src.ml.data.award_training_loader import AwardTrainingLoader, save_dataset_ndjson
-
-        atl = AwardTrainingLoader(taxonomy_version=taxonomy_version or "unknown")
-        # Dispatch based on extension
-        if input_path.suffix.lower() in (".ndjson", ".jsonl"):
-            dataset = atl.load_ndjson(input_path)
-        elif input_path.suffix.lower() == ".csv":
-            dataset = atl.load_csv(input_path)
-        else:
-            # Try CSV then NDJSON as fallback
-            try:
-                dataset = atl.load_csv(input_path)
-            except Exception:
-                dataset = atl.load_ndjson(input_path)
-        stats = atl.last_stats.as_dict() if getattr(atl, "last_stats", None) else {}
-    except Exception as e:
-        # Failed to load training dataset; write checks and empty output
-        try:
-            import pandas as pd  # type: ignore
-
-            df_empty = pd.DataFrame(
-                columns=[
-                    "example_id",
-                    "text",
-                    "title",
-                    "keywords",
-                    "keywords_joined",
-                    "solicitation",
-                    "labels",
-                    "labels_joined",
-                    "source",
-                    "annotated_by",
-                    "annotated_at",
-                    "notes",
-                    "taxonomy_version",
-                ]
-            )
-            save_dataframe_parquet(df_empty, output_path)
-        except Exception:
-            out_json = output_path.with_suffix(".ndjson")
-            out_json.parent.mkdir(parents=True, exist_ok=True)
-            with open(out_json, "w", encoding="utf-8") as fh:
-                fh.write("")
-            try:
-                output_path.parent.mkdir(parents=True, exist_ok=True)
-                output_path.touch()
-            except Exception:
-                pass
-
-        checks = {
-            "ok": False,
-            "reason": "load_failed",
-            "error": str(e),
-            "input_path": str(input_path),
-            "taxonomy_version": taxonomy_version,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "checks_path": str(checks_path),
-            "input_path": str(input_path),
-            "taxonomy_version": taxonomy_version,
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Persist dataset to parquet (preferred) with NDJSON fallback
-    try:
-        import pandas as pd  # type: ignore
-
-        rows = []
-        for ex in dataset.examples:
-            rows.append(
-                {
-                    "example_id": ex.example_id,
-                    "text": ex.text,
-                    "title": ex.title,
-                    "keywords": ex.keywords,
-                    "keywords_joined": (ex.keywords or ""),
-                    "solicitation": ex.solicitation,
-                    "labels": ex.labels,
-                    "labels_joined": ", ".join(ex.labels or []),
-                    "source": ex.source,
-                    "annotated_by": ex.annotated_by,
-                    "annotated_at": ex.annotated_at.isoformat() if ex.annotated_at else None,
-                    "notes": ex.notes,
-                    "taxonomy_version": dataset.taxonomy_version,
-                }
-            )
-        df = pd.DataFrame(rows)
-        save_dataframe_parquet(df, output_path)
-    except Exception:
-        # Fallback to NDJSON using helper
-        try:
-            out_json = output_path.with_suffix(".ndjson")
-            save_dataset_ndjson(dataset, out_json)
-            try:
-                output_path.parent.mkdir(parents=True, exist_ok=True)
-                output_path.touch()
-            except Exception:
-                pass
-        except Exception:
-            # As a last resort, write minimal NDJSON manually
-            out_json = output_path.with_suffix(".ndjson")
-            out_json.parent.mkdir(parents=True, exist_ok=True)
-            with open(out_json, "w", encoding="utf-8") as fh:
-                for ex in dataset.examples:
-                    fh.write(
-                        json.dumps(
-                            {
-                                "example_id": ex.example_id,
-                                "text": ex.text,
-                                "labels": ex.labels,
-                            }
-                        )
-                        + "\n"
-                    )
-            try:
-                output_path.parent.mkdir(parents=True, exist_ok=True)
-                output_path.touch()
-            except Exception:
-                pass
-
-    # Build checks JSON
-    checks = {
-        "ok": True,
-        "rows": len(dataset.examples),
-        "input_path": str(input_path),
-        "taxonomy_version": dataset.taxonomy_version,
-        "stats": stats if isinstance(stats, dict) else {},
-    }
-    checks_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(checks_path, "w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    metadata = {
-        "path": str(output_path),
-        "rows": len(dataset.examples),
-        "checks_path": str(checks_path),
-        "input_path": str(input_path),
-        "taxonomy_version": dataset.taxonomy_version,
-    }
-    return Output(value=str(output_path), metadata=metadata)
-
-
-@asset(
-    name="transformed_cet_analytics",
-    key_prefix=["ml"],
-    description="Compute CET analytics (coverage and specialization) and emit alerts.",
-)
-def transformed_cet_analytics() -> Output:
-    """
-    Compute portfolio-level CET analytics:
-      - Award coverage rate: fraction of awards with a primary CET
-      - Company specialization: average specialization_score across companies
-
-    Emit alerts using AlertCollector when coverage falls below configured threshold.
-    Write a checks JSON under reports/alerts/.
-    """
-    import json
-    from datetime import datetime
-    from pathlib import Path
-
-    try:
-        import pandas as pd  # type: ignore
-    except Exception:
-        pd = None  # type: ignore
-
-    # Lazy import to avoid heavy imports at module import time
-    try:
-        from src.utils.performance_alerts import AlertCollector  # type: ignore
-    except Exception:
-        AlertCollector = None  # type: ignore
-
-    processed_dir = Path("data/processed")
-    alerts_dir = Path("reports/alerts")
-    alerts_dir.mkdir(parents=True, exist_ok=True)
-
-    # Inputs
-    company_parquet = processed_dir / "cet_company_profiles.parquet"
-    company_json = processed_dir / "cet_company_profiles.json"
-    awards_parquet = processed_dir / "cet_award_classifications.parquet"
-    awards_json = processed_dir / "cet_award_classifications.json"
-
-    # Read helpers (parquet preferred, NDJSON fallback)
-    def _read_df(parquet_path: Path, json_path: Path, expected_cols=None):
-        if pd is None:
-            return None
-        if parquet_path.exists():
-            try:
-                df = pd.read_parquet(parquet_path)
-                if expected_cols:
-                    cols = [c for c in expected_cols if c in df.columns]
-                    if cols:
-                        df = df[cols]
-                return df
-            except Exception:
-                pass
-        if json_path.exists():
-            try:
-                rows = []
-                with json_path.open("r", encoding="utf-8") as fh:
-                    for line in fh:
-                        line = line.strip()
-                        if not line:
-                            continue
-                        rows.append(json.loads(line))
-                if not rows:
-                    return pd.DataFrame()
-                df = pd.DataFrame(rows)
-                if expected_cols:
-                    cols = [c for c in expected_cols if c in df.columns]
-                    if cols:
-                        df = df[cols]
-                return df
-            except Exception:
-                return pd.DataFrame()
-        return pd.DataFrame()
-
-    # Load inputs
-    df_companies = _read_df(
-        company_parquet,
-        company_json,
-        expected_cols=[
-            "company_id",
-            "coverage",
-            "specialization_score",
-            "taxonomy_version",
-        ],
-    )
-    df_awards = _read_df(
-        awards_parquet,
-        awards_json,
-        expected_cols=["award_id", "primary_cet", "primary_score", "taxonomy_version"],
-    )
-
-    # Compute metrics (robust to empty frames)
-    coverage_rate = 0.0
-    num_awards = 0
-    num_classified = 0
-    if df_awards is not None and not df_awards.empty:
-        num_awards = len(df_awards)
-        num_classified = (
-            int(df_awards["primary_cet"].notna().sum()) if "primary_cet" in df_awards.columns else 0
-        )
-        coverage_rate = float(num_classified / max(1, num_awards))
-
-    specialization_avg = None
-    if (
-        df_companies is not None
-        and not df_companies.empty
-        and "specialization_score" in df_companies.columns
-    ):
-        specialization_avg = float(df_companies["specialization_score"].dropna().mean())
-
-    # Alerts
-    alerts = {}
-    if AlertCollector is not None:
-        collector = AlertCollector(asset_name="transformed_cet_analytics")
-        # Check coverage_rate against configured match rate threshold
-        collector.check_match_rate(coverage_rate, metric_name="cet_award_coverage_rate")
-        alerts = collector.to_dict()
-        # Persist alerts JSON
-        with (alerts_dir / "cet_analytics.alerts.json").open("w", encoding="utf-8") as fh:
-            json.dump(alerts, fh, indent=2)
-
-    # Checks payload
-    checks = {
-        "ok": True,
-        "generated_at": datetime.utcnow().isoformat(),
-        "award_coverage_rate": coverage_rate,
-        "num_awards": num_awards,
-        "num_classified": num_classified,
-        "company_specialization_avg": specialization_avg,
-        "alerts": alerts,
-    }
-    checks_path = alerts_dir / "cet_analytics.checks.json"
-    try:
-        with checks_path.open("w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-    except Exception:
-        # best-effort
-        pass
-
-    metadata = {
-        "coverage_rate": coverage_rate,
-        "num_awards": num_awards,
-        "num_classified": num_classified,
-        "company_specialization_avg": specialization_avg,
-        "checks_path": str(checks_path),
-        "alerts_path": str(alerts_dir / "cet_analytics.alerts.json"),
-    }
-    return Output(value=metadata, metadata=metadata)
-
-
-@asset(
-    name="raw_cet_human_sampling",
-    key_prefix=["ml"],
-    description="Generate a human-annotation sample from award classifications (balanced by CET where possible).",
-)
-def raw_cet_human_sampling() -> Output:
-    """
-    Produce a small, human-readable sample for annotation.
-
-    - Reads `data/processed/cet_award_classifications.parquet` (preferred) or `.json` NDJSON.
-    - Writes NDJSON sample to `data/processed/cet_human_sample.ndjson`.
-    - Writes a checks JSON to `data/processed/cet_human_sample.checks.json`.
-
-    Sampling is best-effort:
-    - If primary_cet exists, attempt to sample roughly uniformly across CETs.
-    - Otherwise, uniform random sample across all rows.
-    - Configurable via environment variables:
-        - SBIR_ETL__CET__SAMPLE_SIZE (default: 50)
-        - SBIR_ETL__CET__SAMPLE_SEED (default: 42)
-    """
-    import json
-    import os
-    from pathlib import Path
-    from random import Random
-
-    try:
-        import pandas as pd  # type: ignore
-    except Exception:
-        pd = None  # type: ignore
-
-    processed_dir = Path("data/processed")
-    input_parquet = processed_dir / "cet_award_classifications.parquet"
-    input_json = processed_dir / "cet_award_classifications.json"
-
-    output_ndjson = processed_dir / "cet_human_sample.ndjson"
-    checks_path = processed_dir / "cet_human_sample.checks.json"
-
-    # Config
-    sample_size = int(os.environ.get("SBIR_ETL__CET__SAMPLE_SIZE", "50"))
-    seed = int(os.environ.get("SBIR_ETL__CET__SAMPLE_SEED", "42"))
-    Random(seed)
-
-    def _read_awards():
-        if pd is not None and input_parquet.exists():
-            try:
-                return pd.read_parquet(input_parquet)
-            except Exception:
-                pass
-        if input_json.exists():
-            rows = []
-            with input_json.open("r", encoding="utf-8") as fh:
-                for line in fh:
-                    line = line.strip()
-                    if not line:
-                        continue
-                    rows.append(json.loads(line))
-            if rows:
-                try:
-                    return pd.DataFrame(rows)
-                except Exception:
-                    return None
-        return None
-
-    df = _read_awards()
-    total_rows = 0
-    sampled_rows = 0
-
-    if df is None or df.empty:
-        # Write empty sample and checks
-        output_ndjson.write_text("", encoding="utf-8")
-        with checks_path.open("w", encoding="utf-8") as fh:
-            json.dump(
-                {
-                    "ok": True,
-                    "reason": "no_input",
-                    "total_rows": 0,
-                    "sampled_rows": 0,
-                    "source": str(input_parquet if input_parquet.exists() else input_json),
-                },
-                fh,
-                indent=2,
-            )
-        return Output(
-            value=str(output_ndjson),
-            metadata={"path": str(output_ndjson), "rows": 0, "checks_path": str(checks_path)},
-        )
-
-    # Normalize fields for annotation convenience
-    total_rows = len(df)
-    keep_cols = [
-        "award_id",
-        "primary_cet",
-        "primary_score",
-        "supporting_cets",
-        "classified_at",
-        "taxonomy_version",
-        "title",
-        "abstract",
-        "keywords",
-    ]
-    existing_cols = [c for c in keep_cols if c in df.columns]
-    df = df[existing_cols].copy()
-
-    # Sampling strategy: balanced by primary_cet if present, else uniform random
-    if "primary_cet" in df.columns and df["primary_cet"].notna().any():
-        groups = []
-        per_cet = max(1, sample_size // max(1, df["primary_cet"].nunique()))
-        for _cet, sub in df.groupby("primary_cet"):
-            rows = (
-                sub.sample(n=min(per_cet, len(sub)), random_state=seed)
-                if hasattr(sub, "sample")
-                else sub.iloc[:per_cet]
-            )
-            groups.append(rows)
-        df_sample = (
-            pd.concat(groups).drop_duplicates("award_id")
-            if pd is not None
-            else df.iloc[:sample_size]
-        )
-        # If under-filled due to small groups, top-up uniformly
-        if len(df_sample) < sample_size:
-            remaining = df[~df["award_id"].isin(df_sample["award_id"])]
-            top_up = remaining.sample(
-                n=min(sample_size - len(df_sample), len(remaining)), random_state=seed
-            )
-            df_sample = pd.concat([df_sample, top_up]).drop_duplicates("award_id")
-        df_sample = df_sample.head(sample_size)
-    else:
-        # Uniform random
-        if hasattr(df, "sample"):
-            df_sample = df.sample(n=min(sample_size, len(df)), random_state=seed)
-        else:
-            # Fallback deterministic slice
-            df_sample = df.head(sample_size)
-
-    sampled_rows = len(df_sample)
-
-    # Write NDJSON
-    with output_ndjson.open("w", encoding="utf-8") as fh:
-        for _, row in df_sample.iterrows():
-            # Keep only JSON-serializable structures
-            rec = {k: row.get(k) for k in existing_cols}
-            # Ensure keywords and supporting_cets are basic types
-            if isinstance(rec.get("keywords"), list | tuple):
-                rec["keywords"] = list(rec["keywords"])
-            if isinstance(rec.get("supporting_cets"), list | tuple):
-                rec["supporting_cets"] = list(rec["supporting_cets"])
-            fh.write(json.dumps(rec) + "\n")
-
-    # Checks
-    checks = {
-        "ok": True,
-        "total_rows": int(total_rows),
-        "sampled_rows": int(sampled_rows),
-        "balanced_by_primary": "primary_cet" in existing_cols,
-        "seed": seed,
-        "source": str(input_parquet if input_parquet.exists() else input_json),
-    }
-    with checks_path.open("w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    return Output(
-        value=str(output_ndjson),
-        metadata={
-            "path": str(output_ndjson),
-            "rows": int(sampled_rows),
-            "checks_path": str(checks_path),
-        },
-    )
-
-
-@asset(
-    name="validated_cet_iaa_report",
-    key_prefix=["ml"],
-    description="Compute inter-annotator agreement (IAA) for CET labels from human annotations.",
-)
-def validated_cet_iaa_report() -> Output:
-    """
-    Compute inter-annotator agreement (Cohen's kappa and percent agreement) for CET labels.
-
-    Expected input:
-    - Annotation files under `data/processed/annotations/` with extension `.jsonl` or `.ndjson`.
-    - Each line contains at least:
-        { "award_id": "...", "annotator": "userA", "labels": ["cet_id1", "cet_id2", ...] }
-
-    Behavior:
-    - Aligns on award_id across annotators
-    - Converts multi-label sets to a canonical single-label for kappa via:
-        - primary chosen label (first in list) or None
-      and computes:
-        - Cohen's kappa for each annotator pair on the canonical label
-        - Percent agreement on exact set equality across annotators
-    - Writes `reports/analytics/cet_iaa_report.json` and returns a summary.
-    """
-    import json
-    from itertools import combinations
-    from pathlib import Path
-
-    try:
-        import pandas as pd  # type: ignore
-    except Exception:
-        pd = None  # type: ignore
-
-    annotations_dir = Path("data/processed/annotations")
-    reports_dir = Path("reports/analytics")
-    reports_dir.mkdir(parents=True, exist_ok=True)
-    out_path = reports_dir / "cet_iaa_report.json"
-
-    # Collect rows from all .jsonl/.ndjson files
-    rows = []
-    if annotations_dir.exists():
-        for p in annotations_dir.iterdir():
-            if not p.is_file():
-                continue
-            if p.suffix.lower() not in (".jsonl", ".ndjson"):
-                continue
-            try:
-                with p.open("r", encoding="utf-8") as fh:
-                    for line in fh:
-                        line = line.strip()
-                        if not line:
-                            continue
-                        obj = json.loads(line)
-                        rows.append(
-                            {
-                                "award_id": obj.get("award_id"),
-                                "annotator": obj.get("annotator"),
-                                "labels": obj.get("labels") or [],
-                            }
-                        )
-            except Exception:
-                continue
-
-    if not rows or pd is None:
-        payload = {
-            "ok": True,
-            "reason": "no_annotations" if not rows else "pandas_unavailable",
-            "pairs": 0,
-            "kappa": {},
-            "percent_agreement": None,
-            "path": str(out_path),
-        }
-        with out_path.open("w", encoding="utf-8") as fh:
-            json.dump(payload, fh, indent=2)
-        return Output(value=str(out_path), metadata=payload)
-
-    df = pd.DataFrame(rows)
-
-    # Canonical per-annotator primary label (first label or None)
-    def _primary(labels):
-        try:
-            return (labels or [None])[0]
-        except Exception:
-            return None
-
-    df["primary"] = df["labels"].apply(_primary)
-
-    # Build pivot award_id x annotator -> primary label
-    pivot = df.pivot_table(index="award_id", columns="annotator", values="primary", aggfunc="first")
-
-    # Kappa per annotator pair (only on overlapping awards)
-    def _cohen_kappa(series_a, series_b):
-        # Compute Cohen's kappa manually for two categorical series
-        import math
-
-        # Drop pairs with missing values
-        paired = [
-            (a, b) for a, b in zip(series_a, series_b, strict=False) if pd.notna(a) and pd.notna(b)
-        ]
-        if not paired:
-            return None
-        labels = list({x for ab in paired for x in ab})
-        label_to_idx = {label: i for i, label in enumerate(labels)}
-        n = len(paired)
-        # Confusion counts
-        counts = [[0] * len(labels) for _ in labels]
-        for a, b in paired:
-            counts[label_to_idx[a]][label_to_idx[b]] += 1
-        # Observed agreement
-        po = sum(counts[i][i] for i in range(len(labels))) / n
-        # Expected agreement
-        row_marginals = [
-            sum(counts[i][j] for j in range(len(labels))) / n for i in range(len(labels))
-        ]
-        col_marginals = [
-            sum(counts[i][j] for i in range(len(labels))) / n for j in range(len(labels))
-        ]
-        pe = sum(row_marginals[i] * col_marginals[i] for i in range(len(labels)))
-        if math.isclose(1.0 - pe, 0.0):
-            return None
-        return (po - pe) / (1.0 - pe)
-
-    kappa_results = {}
-    annotators = [c for c in pivot.columns if str(c) != "nan"]
-    for a, b in combinations(annotators, 2):
-        s1 = pivot[a].tolist()
-        s2 = pivot[b].tolist()
-        kappa_results[f"{a}__vs__{b}"] = _cohen_kappa(s1, s2)
-
-    # Percent agreement on exact label sets across annotators (only awards with >=2 annotations)
-    set_pivot = df.pivot_table(
-        index="award_id", columns="annotator", values="labels", aggfunc="first"
-    )
-    agree_count = 0
-    denom = 0
-    for _, row in set_pivot.iterrows():
-        non_null = [v for v in row.tolist() if isinstance(v, list)]
-        if len(non_null) < 2:
-            continue
-        denom += 1
-        # Compare all sets for equality
-        eq = True
-        for i in range(1, len(non_null)):
-            if set(non_null[i]) != set(non_null[0]):
-                eq = False
-                break
-        if eq:
-            agree_count += 1
-    percent_agreement = (agree_count / denom) if denom > 0 else None
-
-    payload = {
-        "ok": True,
-        "pairs": int(len(kappa_results)),
-        "kappa": kappa_results,
-        "percent_agreement": percent_agreement,
-        "path": str(out_path),
-    }
-    with out_path.open("w", encoding="utf-8") as fh:
-        json.dump(payload, fh, indent=2)
-
-    return Output(value=str(out_path), metadata=payload)
-
-
-@asset(
-    name="transformed_cet_analytics_aggregates",
-    key_prefix=["ml"],
-    description="Produce CET analytics dashboards (coverage by year and specialization distribution) and alert on regression vs baseline.",
-)
-def transformed_cet_analytics_aggregates() -> Output:
-    """
-    Create CET analytics dashboards and a regression alert vs a baseline:
-      - Coverage by year from cet_award_classifications (derived from classified_at year)
-      - Company specialization distribution from cet_company_profiles
-      - Compare latest-year coverage with a baseline in reports/benchmarks/baseline.json
-      - Write dashboards under reports/analytics and alerts under reports/alerts
-    """
-    import json
-    from datetime import datetime
-    from pathlib import Path
-
-    try:
-        import pandas as pd  # type: ignore
-    except Exception:
-        pd = None  # type: ignore
-
-    # Lazy import to avoid heavy deps at module import time
-    try:
-        from src.utils.performance_alerts import AlertCollector  # type: ignore
-    except Exception:
-        AlertCollector = None  # type: ignore
-
-    processed_dir = Path("data/processed")
-    analytics_dir = Path("reports/analytics")
-    alerts_dir = Path("reports/alerts")
-    baseline_path = Path("reports/benchmarks/baseline.json")
-
-    analytics_dir.mkdir(parents=True, exist_ok=True)
-    alerts_dir.mkdir(parents=True, exist_ok=True)
-
-    # Inputs
-    awards_parquet = processed_dir / "cet_award_classifications.parquet"
-    awards_json = processed_dir / "cet_award_classifications.json"
-    companies_parquet = processed_dir / "cet_company_profiles.parquet"
-    companies_json = processed_dir / "cet_company_profiles.json"
-
-    # Read helpers (parquet preferred, NDJSON fallback)
-    def _read_df(parquet_path: Path, json_path: Path, expected_cols=None):
-        if pd is None:
-            return None
-        if parquet_path.exists():
-            try:
-                df = pd.read_parquet(parquet_path)
-                if expected_cols:
-                    cols = [c for c in expected_cols if c in df.columns]
-                    if cols:
-                        df = df[cols]
-                return df
-            except Exception:
-                pass
-        if json_path.exists():
-            try:
-                rows = []
-                with json_path.open("r", encoding="utf-8") as fh:
-                    for line in fh:
-                        line = line.strip()
-                        if not line:
-                            continue
-                        rows.append(json.loads(line))
-                if not rows:
-                    return pd.DataFrame()
-                df = pd.DataFrame(rows)
-                if expected_cols:
-                    cols = [c for c in expected_cols if c in df.columns]
-                    if cols:
-                        df = df[cols]
-                return df
-            except Exception:
-                return pd.DataFrame()
-        return pd.DataFrame()
-
-    # Load inputs
-    df_awards = _read_df(
-        awards_parquet,
-        awards_json,
-        expected_cols=["award_id", "primary_cet", "classified_at", "taxonomy_version"],
-    )
-    df_companies = _read_df(
-        companies_parquet,
-        companies_json,
-        expected_cols=["company_id", "specialization_score", "taxonomy_version"],
-    )
-
-    # Coverage by year (derive year from classified_at)
-    coverage_by_year = pd.DataFrame()
-    latest_year = None
-    latest_coverage = None
-    total_awards = 0
-    total_classified = 0
-    if df_awards is not None and not df_awards.empty:
-        # Derive year from classified_at; fallback to "unknown"
-        def _year_of(x):
-            try:
-                return str(datetime.fromisoformat(str(x).replace("Z", "+00:00")).year)
-            except Exception:
-                return "unknown"
-
-        df_tmp = df_awards.copy()
-        df_tmp["__year"] = (
-            df_tmp["classified_at"].apply(_year_of)
-            if "classified_at" in df_tmp.columns
-            else "unknown"
-        )
-        grp = df_tmp.groupby("__year", dropna=False)["primary_cet"]
-        coverage_by_year = grp.agg(
-            total_awards=lambda s: int(len(s)),
-            classified=lambda s: int(s.notna().sum()),
-        ).reset_index()
-        coverage_by_year["coverage_rate"] = coverage_by_year["classified"] / coverage_by_year[
-            "total_awards"
-        ].clip(lower=1)
-
-        # Track overall and latest (numeric) year coverage
-        total_awards = int(coverage_by_year["total_awards"].sum())
-        total_classified = int(coverage_by_year["classified"].sum())
-        # Choose latest numeric year for regression comparison
-        try:
-            numeric_years = sorted([int(y) for y in coverage_by_year["__year"] if str(y).isdigit()])
-            if numeric_years:
-                latest_year = numeric_years[-1]
-                latest_row = coverage_by_year[coverage_by_year["__year"] == str(latest_year)].iloc[
-                    0
-                ]
-                latest_coverage = float(latest_row["coverage_rate"])
-        except Exception:
-            latest_year = None
-            latest_coverage = None
-
-    # Company specialization distribution
-    specialization_dist = pd.DataFrame()
-    specialization_avg = None
-    if (
-        df_companies is not None
-        and not df_companies.empty
-        and "specialization_score" in df_companies.columns
-    ):
-        specialization_avg = float(df_companies["specialization_score"].dropna().mean())
-        # Simple histogram buckets
-        bins = [0.0, 0.25, 0.5, 0.75, 1.01]
-        labels = ["[0,0.25)", "[0.25,0.5)", "[0.5,0.75)", "[0.75,1]"]
-        df_tmpc = df_companies.copy()
-        df_tmpc["bucket"] = pd.cut(
-            df_tmpc["specialization_score"].fillna(0.0),
-            bins=bins,
-            labels=labels,
-            include_lowest=True,
-            right=False,
-        )
-        specialization_dist = (
-            df_tmpc.groupby("bucket", dropna=False)["specialization_score"]
-            .agg(count="count")
-            .reset_index()
-        )
-
-    # Write dashboards
-    coverage_csv = analytics_dir / "cet_coverage_by_year.csv"
-    coverage_json = analytics_dir / "cet_coverage_by_year.json"
-    spec_csv = analytics_dir / "cet_company_specialization_distribution.csv"
-    spec_json = analytics_dir / "cet_company_specialization_distribution.json"
-
-    if pd is not None:
-        try:
-            if not coverage_by_year.empty:
-                coverage_by_year.to_csv(coverage_csv, index=False)
-                coverage_by_year.to_json(coverage_json, orient="records", indent=2)
-        except Exception:
-            pass
-        try:
-            if not specialization_dist.empty:
-                specialization_dist.to_csv(spec_csv, index=False)
-                specialization_dist.to_json(spec_json, orient="records", indent=2)
-        except Exception:
-            pass
-
-    # Regression alert vs baseline
-    alerts = {}
-    if AlertCollector is not None:
-        collector = AlertCollector(asset_name="transformed_cet_analytics_aggregates")
-        baseline_min = None
-        try:
-            if baseline_path.exists():
-                with baseline_path.open("r", encoding="utf-8") as fh:
-                    baseline = json.load(fh)
-                # Support a couple of common shapes
-                # e.g., {"cet": {"coverage_min": 0.6}} or {"coverage_min": 0.6}
-                if isinstance(baseline, dict):
-                    if "cet" in baseline and isinstance(baseline["cet"], dict):
-                        baseline_min = float(baseline["cet"].get("coverage_min", 0.0))
-                    elif "coverage_min" in baseline:
-                        baseline_min = float(baseline.get("coverage_min", 0.0))
-        except Exception:
-            baseline_min = None
-
-        # If baseline exists and latest_coverage available, compare
-        if baseline_min is not None and latest_coverage is not None:
-            # Use check_match_rate semantics (alerts FAILURE if below threshold)
-            collector.check_match_rate(
-                latest_coverage, metric_name="cet_award_latest_year_coverage"
-            )
-            # If baseline_min higher than coverage, force metadata
-            if latest_coverage < baseline_min:
-                # Already represented as FAILURE by check_match_rate when threshold configured accordingly.
-                pass
-        alerts = collector.to_dict()
-        try:
-            with (alerts_dir / "cet_analytics_aggregates.alerts.json").open(
-                "w", encoding="utf-8"
-            ) as fh:
-                json.dump(alerts, fh, indent=2)
-        except Exception:
-            pass
-
-    metadata = {
-        "coverage_dashboard_csv": str(coverage_csv),
-        "coverage_dashboard_json": str(coverage_json),
-        "specialization_dashboard_csv": str(spec_csv),
-        "specialization_dashboard_json": str(spec_json),
-        "latest_year": latest_year,
-        "latest_coverage_rate": latest_coverage,
-        "total_awards": total_awards,
-        "total_classified": total_classified,
-        "specialization_avg": specialization_avg,
-        "alerts_path": str(alerts_dir / "cet_analytics_aggregates.alerts.json"),
-    }
-    return Output(value=metadata, metadata=metadata)
-
-
-@asset(
-    name="validated_cet_drift_detection",
-    key_prefix=["ml"],
-    description=(
-        "Detect distributional drift for CET classifications and model scores by "
-        "comparing current distributions to a stored baseline. Emits alerts and "
-        "writes a small report to `reports/alerts` and `reports/benchmarks`."
-    ),
-)
-def validated_cet_drift_detection() -> Output:
-    """
-    Model drift detection asset for CET classification outputs.
-
-    Behavior (best-effort / import-safe):
-    - Loads award-level CET classifications from `data/processed/cet_award_classifications.parquet`
-      or NDJSON fallback.
-    - Computes two simple distributions:
-        * primary_score histogram (continuous scores)
-        * primary_cet frequency distribution (categorical)
-    - Attempts to load a baseline distributions file at `reports/benchmarks/cet_baseline_distributions.json`.
-      If baseline is missing, the asset writes current distributions as a baseline candidate
-      to `reports/benchmarks/cet_baseline_distributions_current.json` and returns with no alert.
-    - If baseline exists, computes Jensen-Shannon divergence (symmetric) between baseline and current
-      distributions for both score and label distributions. If divergence exceeds thresholds (configurable
-      via env vars), writes an alerts JSON and emits a non-fatal report.
-    - Writes:
-        - reports/benchmarks/cet_drift_report.json  (summary & divergence values)
-        - reports/alerts/cet_drift_alerts.json       (alerts if any)
-    """
-    import json
-    import os
-    from datetime import datetime
-    from pathlib import Path
-
-    try:
-        import numpy as np  # type: ignore
-        import pandas as pd  # type: ignore
-    except Exception:
-        pd = None  # type: ignore
-        np = None  # type: ignore
-
-    # Lazy import for AlertCollector (best-effort)
-    try:
-        from src.utils.performance_alerts import (  # type: ignore
-            Alert,
-            AlertCollector,
-            AlertSeverity,
-        )
-    except Exception:
-        AlertCollector = None  # type: ignore
-        Alert = None  # type: ignore
-        AlertSeverity = None  # type: ignore
-
-    processed_dir = Path("data/processed")
-    awards_parquet = processed_dir / "cet_award_classifications.parquet"
-    awards_json = processed_dir / "cet_award_classifications.json"
-
-    benchmarks_dir = Path("reports/benchmarks")
-    benchmarks_dir.mkdir(parents=True, exist_ok=True)
-    alerts_dir = Path("reports/alerts")
-    alerts_dir.mkdir(parents=True, exist_ok=True)
-
-    baseline_path = benchmarks_dir / "cet_baseline_distributions.json"
-    baseline_candidate_path = benchmarks_dir / "cet_baseline_distributions_current.json"
-    drift_report = benchmarks_dir / "cet_drift_report.json"
-    alerts_out = alerts_dir / "cet_drift_alerts.json"
-
-    # Configurable thresholds (env vars, fallback defaults)
-    SCORE_JS_THRESHOLD = float(os.environ.get("SBIR_ETL__CET__DRIFT__SCORE_JS_THRESHOLD", "0.15"))
-    LABEL_JS_THRESHOLD = float(os.environ.get("SBIR_ETL__CET__DRIFT__LABEL_JS_THRESHOLD", "0.10"))
-
-    # Helper: safe JSON read
-    def _read_json(path: Path):
-        try:
-            if path.exists():
-                with path.open("r", encoding="utf-8") as fh:
-                    return json.load(fh)
-        except Exception:
-            return None
-        return None
-
-    # Read awards
-    def _read_awards():
-        if pd is None:
-            return None
-        if awards_parquet.exists():
-            try:
-                return pd.read_parquet(awards_parquet)
-            except Exception:
-                pass
-        if awards_json.exists():
-            try:
-                rows = []
-                with awards_json.open("r", encoding="utf-8") as fh:
-                    for line in fh:
-                        line = line.strip()
-                        if not line:
-                            continue
-                        rows.append(json.loads(line))
-                return pd.DataFrame(rows)
-            except Exception:
-                return pd.DataFrame()
-        return pd.DataFrame()
-
-    df = _read_awards()
-    if df is None or df.empty:
-        # Nothing to do; write an empty report and return
-        payload = {
-            "ok": True,
-            "reason": "no_input",
-            "generated_at": datetime.utcnow().isoformat(),
-            "score_js_divergence": None,
-            "label_js_divergence": None,
-            "score_threshold": SCORE_JS_THRESHOLD,
-            "label_threshold": LABEL_JS_THRESHOLD,
-        }
-        try:
-            with drift_report.open("w", encoding="utf-8") as fh:
-                json.dump(payload, fh, indent=2)
-        except Exception:
-            pass
-        return Output(value=payload, metadata=payload)
-
-    # Prepare current distributions
-    # Primary scores (coerce to floats, dropna)
-    scores = None
-    if "primary_score" in df.columns:
-        try:
-            scores = pd.to_numeric(df["primary_score"], errors="coerce").dropna().astype(float)
-        except Exception:
-            scores = None
-
-    # Primary CET frequency
-    label_counts = {}
-    if "primary_cet" in df.columns:
-        try:
-            label_counts = df["primary_cet"].fillna("__none__").value_counts().to_dict()
-        except Exception:
-            label_counts = {}
-
-    # Convert counts to probability mass function (PMF) for labels
-    def _pmf_from_counts(counts: dict):
-        total = sum(v for v in counts.values() if v is not None)
-        if total <= 0:
-            return {}
-        return {k: float(v) / total for k, v in counts.items()}
-
-    current_label_pmf = _pmf_from_counts(label_counts)
-
-    # For score histogram, create fixed bins (0..100 by 10)
-    def _score_hist_pmf(series, bins=None):
-        if series is None or len(series) == 0 or np is None:
-            return {}
-        if bins is None:
-            bins = list(range(0, 101, 10))  # 0-10,10-20,...,90-100
-        try:
-            hist, bin_edges = np.histogram(series.clip(0, 100), bins=bins)
-            total = int(hist.sum())
-            if total == 0:
-                return {}
-            pmf = {}
-            for i in range(len(hist)):
-                label = f"{int(bin_edges[i])}-{int(bin_edges[i+1])}"
-                pmf[label] = float(hist[i]) / total
-            return pmf
-        except Exception:
-            return {}
-
-    current_score_pmf = _score_hist_pmf(scores)
-
-    # If no baseline exists, write candidate and exit (operator can promote to baseline manually)
-    baseline = _read_json(baseline_path)
-    if baseline is None:
-        # Write current distributions as candidate baseline
-        candidate = {
-            "generated_at": datetime.utcnow().isoformat(),
-            "label_pmf": current_label_pmf,
-            "score_pmf": current_score_pmf,
-        }
-        try:
-            with baseline_candidate_path.open("w", encoding="utf-8") as fh:
-                json.dump(candidate, fh, indent=2)
-        except Exception:
-            pass
-        payload = {
-            "ok": True,
-            "reason": "baseline_missing",
-            "message": "Baseline distributions not found; wrote current distributions as candidate",
-            "candidate_path": str(baseline_candidate_path),
-            "generated_at": datetime.utcnow().isoformat(),
-        }
-        try:
-            with drift_report.open("w", encoding="utf-8") as fh:
-                json.dump(payload, fh, indent=2)
-        except Exception:
-            pass
-        return Output(value=payload, metadata=payload)
-
-    # Baseline exists: extract baseline pmfs (expected keys: label_pmf, score_pmf)
-    baseline_label_pmf = baseline.get("label_pmf", {}) if isinstance(baseline, dict) else {}
-    baseline_score_pmf = baseline.get("score_pmf", {}) if isinstance(baseline, dict) else {}
-
-    # Ensure support alignment for label pmfs (add zeros for missing labels)
-    def _align_pmfs(p, q):
-        keys = set(p.keys()) | set(q.keys())
-        p_arr = [p.get(k, 0.0) for k in sorted(keys)]
-        q_arr = [q.get(k, 0.0) for k in sorted(keys)]
-        labels = sorted(keys)
-        return p_arr, q_arr, labels
-
-    # Jensen-Shannon divergence (symmetric, bounded 0..1)
-    def _js_divergence(p_arr, q_arr):
-        try:
-            import math
-        except Exception:
-            # Fallback simple impl using pure math
-            pass
-        # Convert to numpy arrays if available for numeric stability
-        if np is not None:
-            p = np.array(p_arr, dtype=float)
-            q = np.array(q_arr, dtype=float)
-            # normalize
-            p_sum = p.sum()
-            q_sum = q.sum()
-            if p_sum > 0:
-                p = p / p_sum
-            if q_sum > 0:
-                q = q / q_sum
-            m = 0.5 * (p + q)
-
-            # Use KL divergence helper with safe handling
-            def _kl(a, b):
-                # kl divergence sum a * log2(a/b) where 0*log(0)=0
-                mask = (a > 0) & (b > 0)
-                if mask.sum() == 0:
-                    return 0.0
-                return float(np.sum(a[mask] * np.log2(a[mask] / b[mask])))
-
-            kl_pm = _kl(p, m)
-            kl_qm = _kl(q, m)
-            js = 0.5 * (kl_pm + kl_qm)
-            # Convert to 0..1 by dividing by log2(len) if len>1 (optional). Keep raw JS for thresholding.
-            return float(js)
-        else:
-            # Pure python fallback
-            import math
-
-            def safe_log2(x):
-                return math.log(x, 2) if x > 0 else 0.0
-
-            p = list(map(float, p_arr))
-            q = list(map(float, q_arr))
-            p_sum = sum(p)
-            q_sum = sum(q)
-            if p_sum > 0:
-                p = [x / p_sum for x in p]
-            if q_sum > 0:
-                q = [x / q_sum for x in q]
-            m = [0.5 * (pp + qq) for pp, qq in zip(p, q, strict=False)]
-
-            def kl(a, b):
-                s = 0.0
-                for ai, bi in zip(a, b, strict=False):
-                    if ai > 0 and bi > 0:
-                        s += ai * safe_log2(ai / bi)
-                return s
-
-            return 0.5 * (kl(p, m) + kl(q, m))
-
-    # Align and compute divergences
-    p_label, q_label, label_keys = _align_pmfs(baseline_label_pmf, current_label_pmf)
-    label_js = _js_divergence(p_label, q_label) if label_keys else None
-
-    p_score, q_score, score_keys = _align_pmfs(baseline_score_pmf, current_score_pmf)
-    score_js = _js_divergence(p_score, q_score) if score_keys else None
-
-    # Compose report
-    report = {
-        "generated_at": datetime.utcnow().isoformat(),
-        "label_js_divergence": label_js,
-        "score_js_divergence": score_js,
-        "label_keys": label_keys,
-        "score_keys": score_keys,
-        "score_threshold": SCORE_JS_THRESHOLD,
-        "label_threshold": LABEL_JS_THRESHOLD,
-    }
-
-    # Build alerts based on thresholds
-    alerts_payload = {"alerts": [], "generated_at": datetime.utcnow().isoformat()}
-    if label_js is not None and label_js > LABEL_JS_THRESHOLD:
-        alerts_payload["alerts"].append(
-            {
-                "severity": "WARNING" if label_js <= 2 * LABEL_JS_THRESHOLD else "FAILURE",
-                "type": "label_distribution_drift",
-                "message": f"Label distribution JS divergence {label_js:.4f} exceeds threshold {LABEL_JS_THRESHOLD:.4f}",
-                "value": label_js,
-            }
-        )
-    if score_js is not None and score_js > SCORE_JS_THRESHOLD:
-        alerts_payload["alerts"].append(
-            {
-                "severity": "WARNING" if score_js <= 2 * SCORE_JS_THRESHOLD else "FAILURE",
-                "type": "score_distribution_drift",
-                "message": f"Score distribution JS divergence {score_js:.4f} exceeds threshold {SCORE_JS_THRESHOLD:.4f}",
-                "value": score_js,
-            }
-        )
-
-    # Persist report and alerts
-    try:
-        with drift_report.open("w", encoding="utf-8") as fh:
-            json.dump(report, fh, indent=2)
-    except Exception:
-        pass
-
-    try:
-        with alerts_out.open("w", encoding="utf-8") as fh:
-            json.dump(alerts_payload, fh, indent=2)
-    except Exception:
-        pass
-
-    # Optionally log alerts via AlertCollector if available (best-effort)
-    if AlertCollector is not None and Alert is not None and AlertSeverity is not None:
-        try:
-            collector = AlertCollector(asset_name="validated_cet_drift_detection")
-            for a in alerts_payload.get("alerts", []):
-                sev = AlertSeverity.WARNING if a["severity"] == "WARNING" else AlertSeverity.FAILURE
-                alert = Alert(
-                    timestamp=datetime.utcnow(),
-                    severity=sev,
-                    alert_type=a["type"],
-                    message=a["message"],
-                    threshold_value=LABEL_JS_THRESHOLD
-                    if "label" in a["type"]
-                    else SCORE_JS_THRESHOLD,
-                    actual_value=a["value"],
-                    metric_name=a["type"],
-                )
-                collector.alerts.append(alert)
-            # Save structured alerts JSON via collector
-            try:
-                collector.save_to_file(alerts_out)
-            except Exception:
-                pass
-        except Exception:
-            pass
-
-    # Return the report as asset output
-    metadata = {"drift_report": str(drift_report), "alerts": str(alerts_out)}
-    return Output(value=report, metadata=metadata)
-
-
-@asset(
-    name="transformed_cet_company_profiles",
-    key_prefix=["ml"],
-    description=(
-        "Aggregate award-level CET classifications into company-level CET profiles, "
-        "persist results to `data/processed/cet_company_profiles.parquet` (parquet -> NDJSON "
-        "fallback) and emit a companion checks JSON for automated validation."
-    ),
-)
-def transformed_cet_company_profiles() -> Output:
-    """
-    Dagster asset to perform company-level aggregation of CET classifications.
-
-    Behavior (best-effort / import-safe):
-    - Attempts to load `data/processed/cet_award_classifications.parquet` or `.json` NDJSON fallback.
-      If the classifications input is missing, produces an empty company profiles output so downstream
-      consumers have a deterministic schema.
-    - Uses `CompanyCETAggregator` (from `src.transformers.company_cet_aggregator`) to compute per-company
-      CET aggregates: coverage, dominant CET, specialization (HHI), CET score map, and trend.
-    - Persists company profiles to `data/processed/cet_company_profiles.parquet` with NDJSON fallback.
-    - Writes a checks JSON summarizing company count and basic coverage metrics.
-    """
-    logger.info("Starting cet_company_profiles asset")
-
-    # Local imports to keep module import-safe when optional deps are missing
-    import json
-    from pathlib import Path
-
-    try:
-        import pandas as pd
-    except Exception:
-        pd = None  # type: ignore
-
-    try:
-        from src.transformers.company_cet_aggregator import CompanyCETAggregator
-    except Exception:
-        CompanyCETAggregator = None  # type: ignore
-
-    # Paths
-    classifications_parquet = Path("data/processed/cet_award_classifications.parquet")
-    classifications_ndjson = Path("data/processed/cet_award_classifications.json")
-    output_path = Path("data/processed/cet_company_profiles.parquet")
-    checks_path = output_path.with_suffix(".checks.json")
-
-    # If dependencies missing, write placeholder output & checks
-    if pd is None or CompanyCETAggregator is None:
-        logger.warning(
-            "Missing dependencies for company aggregation (pandas: %s, aggregator: %s). Writing placeholder output.",
-            pd is not None,
-            CompanyCETAggregator is not None,
-        )
-        # Produce an empty DataFrame with expected columns so downstream consumers have schema
-        if pd is not None:
-            df_empty = pd.DataFrame(
-                columns=[
-                    "company_id",
-                    "company_name",
-                    "total_awards",
-                    "awards_with_cet",
-                    "coverage",
-                    "dominant_cet",
-                    "dominant_score",
-                    "specialization_score",
-                    "cet_scores",
-                    "first_award_date",
-                    "last_award_date",
-                    "cet_trend",
-                ]
-            )
-            try:
-                save_dataframe_parquet(df_empty, output_path)
-            except Exception:
-                out_json = output_path.with_suffix(".json")
-                out_json.parent.mkdir(parents=True, exist_ok=True)
-                with open(out_json, "w", encoding="utf-8") as fh:
-                    fh.write("")
-        checks = {
-            "ok": False,
-            "reason": "missing_dependency",
-            "pandas_present": pd is not None,
-            "aggregator_present": CompanyCETAggregator is not None,
-        }
-        checks_path.parent.mkdir(parents=True, exist_ok=True)
-        with open(checks_path, "w", encoding="utf-8") as fh:
-            json.dump(checks, fh, indent=2)
-        metadata = {
-            "path": str(output_path),
-            "rows": 0,
-            "checks_path": str(checks_path),
-        }
-        return Output(value=str(output_path), metadata=metadata)
-
-    # Load classifications (prefer parquet, then NDJSON)
-    try:
-        if classifications_parquet.exists():
-            df_cls = pd.read_parquet(classifications_parquet)
-        elif classifications_ndjson.exists():
-            recs = []
-            with open(classifications_ndjson, encoding="utf-8") as fh:
-                for line in fh:
-                    if line.strip():
-                        recs.append(json.loads(line))
-            df_cls = pd.DataFrame(recs)
-        else:
-            logger.warning(
-                "No cet_award_classifications found at expected paths; producing empty company profiles"
-            )
-            df_cls = pd.DataFrame(
-                columns=[
-                    "award_id",
-                    "company_id",
-                    "company_name",
-                    "primary_cet",
-                    "primary_score",
-                    "supporting_cets",
-                    "classified_at",
-                    "award_date",
-                    "phase",
-                ]
-            )
-    except Exception:
-        logger.exception("Failed to load award classifications; producing empty company profiles")
-        df_cls = pd.DataFrame(
-            columns=[
-                "award_id",
-                "company_id",
-                "company_name",
-                "primary_cet",
-                "primary_score",
-                "supporting_cets",
-                "classified_at",
-                "award_date",
-                "phase",
-            ]
-        )
-
-    # Join with enriched awards to get company information if missing
-    if not df_cls.empty and "company_id" not in df_cls.columns:
-        try:
-            enriched_awards_parquet = Path("data/processed/enriched_sbir_awards.parquet")
-            enriched_awards_ndjson = Path("data/processed/enriched_sbir_awards.ndjson")
-            if enriched_awards_parquet.exists():
-                df_awards = pd.read_parquet(enriched_awards_parquet)
-            elif enriched_awards_ndjson.exists():
-                recs = []
-                with open(enriched_awards_ndjson, encoding="utf-8") as fh:
-                    for line in fh:
-                        if line.strip():
-                            recs.append(json.loads(line))
-                df_awards = pd.DataFrame(recs)
-            else:
-                df_awards = pd.DataFrame()
-
-            if not df_awards.empty and "award_id" in df_cls.columns:
-                # Try to find award ID column in enriched awards
-                award_id_col = None
-                for col in ["award_id", "Award ID", "id"]:
-                    if col in df_awards.columns:
-                        award_id_col = col
-                        break
-
-                if award_id_col:
-                    # Try to find company identifier columns
-                    company_id_col = None
-                    company_name_col = None
-                    for col in ["company_uei", "UEI", "company_id"]:
-                        if col in df_awards.columns:
-                            company_id_col = col
-                            break
-                    for col in ["Company", "company_name", "company"]:
-                        if col in df_awards.columns:
-                            company_name_col = col
-                            break
-
-                    # Join on award_id
-                    if company_id_col:
-                        join_cols = [award_id_col, company_id_col]
-                        if company_name_col and company_name_col not in join_cols:
-                            join_cols.append(company_name_col)
-                        df_join = df_awards[
-                            [col for col in join_cols if col in df_awards.columns]
-                        ].copy()
-                        df_join = df_join.rename(
-                            columns={award_id_col: "award_id", company_id_col: "company_id"}
-                        )
-                        if company_name_col and company_name_col != "company_id":
-                            df_join = df_join.rename(columns={company_name_col: "company_name"})
-                        df_cls = df_cls.merge(df_join, on="award_id", how="left")
-                        logger.info(
-                            f"Joined classifications with enriched awards to get company info (joined {df_cls['company_id'].notna().sum()} rows)"
-                        )
-        except Exception:
-            logger.exception(
-                "Failed to join classifications with enriched awards; proceeding without company info"
-            )
-
-    # Ensure company_id and company_name columns exist (CompanyCETAggregator expects them)
-    if not df_cls.empty:
-        if "company_id" not in df_cls.columns:
-            df_cls["company_id"] = None
-        if "company_name" not in df_cls.columns:
-            df_cls["company_name"] = None
-
-    # Run aggregation
-    try:
-        aggregator = CompanyCETAggregator(df_cls)
-        df_comp = aggregator.to_dataframe()
-    except Exception:
-        logger.exception("Company aggregation failed; producing empty company profiles")
-        df_comp = pd.DataFrame(
-            columns=[
-                "company_id",
-                "company_name",
-                "total_awards",
-                "awards_with_cet",
-                "coverage",
-                "dominant_cet",
-                "dominant_score",
-                "specialization_score",
-                "cet_scores",
-                "first_award_date",
-                "last_award_date",
-                "cet_trend",
-            ]
-        )
-
-    # Persist company profiles (parquet preferred, NDJSON fallback)
-    try:
-        save_dataframe_parquet(df_comp, output_path)
-    except Exception:
-        # Fallback: write NDJSON manually
-        json_out = output_path.with_suffix(".json")
-        json_out.parent.mkdir(parents=True, exist_ok=True)
-        with open(json_out, "w", encoding="utf-8") as fh:
-            for rec in df_comp.to_dict(orient="records"):
-                fh.write(json.dumps(rec) + "\n")
-        # Touch parquet placeholder for consumers that assert its existence
-        try:
-            output_path.parent.mkdir(parents=True, exist_ok=True)
-            output_path.touch()
-        except Exception:
-            logger.exception("Failed to touch parquet placeholder file for company profiles")
-
-    # Build checks
-    num_companies = len(df_comp)
-    checks = {
-        "ok": True,
-        "num_companies": int(num_companies),
-        "num_records_written": int(num_companies),
-    }
-    checks_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(checks_path, "w", encoding="utf-8") as fh:
-        json.dump(checks, fh, indent=2)
-
-    metadata = {
-        "path": str(output_path),
-        "rows": len(df_comp),
-        "checks_path": str(checks_path),
-    }
-
-    logger.info("Completed cet_company_profiles asset", rows=len(df_comp), output=str(output_path))
-
-    return Output(value=str(output_path), metadata=metadata)
-
-
-# ============================================================================
-# Neo4j Loading Assets (Consolidated from cet_neo4j_loading_assets.py)
-# ============================================================================
-
-# Neo4j loader imports (import-safe)
-try:
-    from src.loaders.neo4j_client import LoadMetrics, Neo4jClient, Neo4jConfig  # type: ignore
-except Exception:  # pragma: no cover
-    Neo4jClient = None  # type: ignore
-    Neo4jConfig = None  # type: ignore
-    LoadMetrics = None  # type: ignore
-
-try:
-    from src.loaders.cet_loader import CETLoader, CETLoaderConfig  # type: ignore
-except Exception:  # pragma: no cover
-    CETLoader = None  # type: ignore
-    CETLoaderConfig = None  # type: ignore
-
-# Configuration Defaults for Neo4j Loading
-DEFAULT_NEO4J_URI = os.environ.get("NEO4J_URI", "bolt://localhost:7687")
-DEFAULT_NEO4J_USER = os.environ.get("NEO4J_USER", "neo4j")
-DEFAULT_NEO4J_PASSWORD = os.environ.get("NEO4J_PASSWORD", "neo4j")
-DEFAULT_NEO4J_DATABASE = os.environ.get("NEO4J_DATABASE", "neo4j")
-
-DEFAULT_PROCESSED_DIR_NEO4J = Path("data/processed")
-DEFAULT_TAXONOMY_PARQUET = DEFAULT_PROCESSED_DIR_NEO4J / "cet_taxonomy.parquet"
-DEFAULT_TAXONOMY_JSON = DEFAULT_PROCESSED_DIR_NEO4J / "cet_taxonomy.json"
-
-DEFAULT_AWARD_CLASS_PARQUET = DEFAULT_PROCESSED_DIR_NEO4J / "cet_award_classifications.parquet"
-DEFAULT_AWARD_CLASS_JSON = DEFAULT_PROCESSED_DIR_NEO4J / "cet_award_classifications.json"
-
-DEFAULT_COMPANY_PROFILES_PARQUET = DEFAULT_PROCESSED_DIR_NEO4J / "cet_company_profiles.parquet"
-DEFAULT_COMPANY_PROFILES_JSON = DEFAULT_PROCESSED_DIR_NEO4J / "cet_company_profiles.json"
-
-DEFAULT_OUTPUT_DIR = Path(os.environ.get("SBIR_ETL__CET__NEO4J_OUTPUT_DIR", "data/loaded/neo4j"))
-
-
-def _get_neo4j_client():
-    """Get Neo4j client with error handling."""
-    if Neo4jClient is None or Neo4jConfig is None:
-        return None
-    try:
-        config = Neo4jConfig(
-            uri=DEFAULT_NEO4J_URI,
-            user=DEFAULT_NEO4J_USER,
-            password=DEFAULT_NEO4J_PASSWORD,
-            database=DEFAULT_NEO4J_DATABASE,
-        )
-        return Neo4jClient(config)
-    except Exception:
-        return None
-
-
-def _read_parquet_or_ndjson(
-    parquet_path: Path, json_path: Path, expected_columns: tuple
-) -> list[dict]:
-    """Read data from parquet or fallback to NDJSON."""
-    if pd is None:
-        return []
-
-    try:
-        if parquet_path.exists():
-            df = pd.read_parquet(parquet_path)
-            return df.to_dict(orient="records")
-        elif json_path.exists():
-            records = []
-            with json_path.open("r", encoding="utf-8") as fh:
-                for line in fh:
-                    if line.strip():
-                        try:
-                            records.append(json.loads(line))
-                        except Exception:
-                            continue
-            return records
-    except Exception:
-        pass
-    return []
-
-
-def _serialize_metrics(metrics) -> dict[str, Any]:
-    """Serialize LoadMetrics to dict."""
-    if metrics is None:
-        return {}
-    return {
-        "nodes_created": getattr(metrics, "nodes_created", 0),
-        "nodes_updated": getattr(metrics, "nodes_updated", 0),
-        "relationships_created": getattr(metrics, "relationships_created", 0),
-        "relationships_updated": getattr(metrics, "relationships_updated", 0),
-        "execution_time_ms": getattr(metrics, "execution_time_ms", 0),
-    }
-
-
-@asset(
-    name="loaded_cet_areas",
-    description="Load CETArea nodes into Neo4j from CET taxonomy artifact.",
-    group_name="neo4j_cet",
-    ins={"cet_taxonomy": AssetIn(key=["ml", "raw_cet_taxonomy"])},
-    config_schema={
-        "create_constraints": bool,
-        "create_indexes": bool,
-        "taxonomy_parquet": str,
-        "taxonomy_json": str,
-        "batch_size": int,
-    },
-)
-def loaded_cet_areas(context, cet_taxonomy) -> dict[str, Any]:
-    """Upsert CETArea nodes based on taxonomy output."""
-    if CETLoader is None or CETLoaderConfig is None:
-        context.log.warning("CETLoader unavailable; skipping CETArea loading")
-        return {"status": "skipped", "reason": "loader_unavailable"}
-
-    client = _get_neo4j_client()
-    if client is None:
-        return {"status": "skipped", "reason": "neo4j_unavailable"}
-
-    # Config
-    taxonomy_parquet = Path(
-        context.op_config.get("taxonomy_parquet") or str(DEFAULT_TAXONOMY_PARQUET)
-    )
-    taxonomy_json = Path(context.op_config.get("taxonomy_json") or str(DEFAULT_TAXONOMY_JSON))
-    create_constraints = bool(context.op_config.get("create_constraints", True))
-    create_indexes = bool(context.op_config.get("create_indexes", True))
-    batch_size = int(context.op_config.get("batch_size", 1000))
-
-    # Read taxonomy (expect: cet_id, name, definition, keywords, taxonomy_version)
-    expected_cols = ("cet_id", "name", "definition", "keywords", "taxonomy_version")
-    areas = _read_parquet_or_ndjson(taxonomy_parquet, taxonomy_json, expected_columns=expected_cols)
-    context.log.info(f"Loaded CET taxonomy records for Neo4j: {len(areas)}")
-
-    try:
-        loader = CETLoader(client, CETLoaderConfig(batch_size=batch_size))
-        if create_constraints:
-            loader.create_constraints()
-        if create_indexes:
-            loader.create_indexes()
-
-        metrics = loader.load_cet_areas(areas)
-        result = {
-            "status": "success",
-            "areas": len(areas),
-            "metrics": _serialize_metrics(metrics),
-        }
-
-        # Persist a small run summary
-        out_path = DEFAULT_OUTPUT_DIR / "neo4j_cetarea_nodes.checks.json"
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            with out_path.open("w", encoding="utf-8") as fh:
-                json.dump(result, fh, indent=2)
-        except Exception:
-            pass
-
-        try:
-            client.close()
-        except Exception:
-            pass
-        return result
-    except Exception as exc:
-        context.log.exception(f"CETArea loading failed: {exc}")
-        try:
-            client.close()
-        except Exception:
-            pass
-        return {"status": "error", "error": str(exc)}
-
-
-@asset(
-    name="loaded_award_cet_enrichment",
-    description="Upsert CET enrichment properties onto Award nodes from award classifications artifact.",
-    group_name="neo4j_cet",
-    ins={
-        "enriched_cet_award_classifications": AssetIn(
-            key=["ml", "enriched_cet_award_classifications"]
-        ),
-        "loaded_cet_areas": AssetIn(),
-    },
-    config_schema={
-        "award_class_parquet": str,
-        "award_class_json": str,
-        "batch_size": int,
-    },
-)
-def loaded_award_cet_enrichment(
-    context, enriched_cet_award_classifications, loaded_cet_areas
-) -> dict[str, Any]:
-    """Upsert CET enrichment properties onto Award nodes."""
-    if CETLoader is None or CETLoaderConfig is None:
-        context.log.warning("CETLoader unavailable; skipping Award CET enrichment")
-        return {"status": "skipped", "reason": "loader_unavailable"}
-
-    client = _get_neo4j_client()
-    if client is None:
-        return {"status": "skipped", "reason": "neo4j_unavailable"}
-
-    # Config
-    award_class_parquet = Path(
-        context.op_config.get("award_class_parquet") or str(DEFAULT_AWARD_CLASS_PARQUET)
-    )
-    award_class_json = Path(
-        context.op_config.get("award_class_json") or str(DEFAULT_AWARD_CLASS_JSON)
-    )
-    batch_size = int(context.op_config.get("batch_size", 1000))
-
-    # Read award classifications
-    expected_cols = ("award_id", "primary_cet", "supporting_cets", "confidence", "evidence")
-    classifications = _read_parquet_or_ndjson(
-        award_class_parquet, award_class_json, expected_columns=expected_cols
-    )
-    context.log.info(f"Loaded award classifications for Neo4j: {len(classifications)}")
-
-    try:
-        loader = CETLoader(client, CETLoaderConfig(batch_size=batch_size))
-        metrics = loader.load_award_cet_enrichment(classifications)
-        result = {
-            "status": "success",
-            "awards": len(classifications),
-            "metrics": _serialize_metrics(metrics),
-        }
-
-        # Persist a small run summary
-        out_path = DEFAULT_OUTPUT_DIR / "neo4j_award_cet_enrichment.checks.json"
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            with out_path.open("w", encoding="utf-8") as fh:
-                json.dump(result, fh, indent=2)
-        except Exception:
-            pass
-
-        try:
-            client.close()
-        except Exception:
-            pass
-        return result
-    except Exception as exc:
-        context.log.exception(f"Award CET enrichment failed: {exc}")
-        try:
-            client.close()
-        except Exception:
-            pass
-        return {"status": "error", "error": str(exc)}
-
-
-@asset(
-    name="loaded_company_cet_enrichment",
-    description="Upsert CET enrichment properties onto Company nodes from company CET profiles.",
-    group_name="neo4j_cet",
-    ins={
-        "transformed_cet_company_profiles": AssetIn(key=["ml", "transformed_cet_company_profiles"]),
-        "loaded_cet_areas": AssetIn(),
-    },
-    config_schema={
-        "company_profiles_parquet": str,
-        "company_profiles_json": str,
-        "batch_size": int,
-    },
-)
-def loaded_company_cet_enrichment(
-    context, transformed_cet_company_profiles, loaded_cet_areas
-) -> dict[str, Any]:
-    """Upsert CET enrichment properties onto Company nodes."""
-    if CETLoader is None or CETLoaderConfig is None:
-        context.log.warning("CETLoader unavailable; skipping Company CET enrichment")
-        return {"status": "skipped", "reason": "loader_unavailable"}
-
-    client = _get_neo4j_client()
-    if client is None:
-        return {"status": "skipped", "reason": "neo4j_unavailable"}
-
-    # Config
-    company_profiles_parquet = Path(
-        context.op_config.get("company_profiles_parquet") or str(DEFAULT_COMPANY_PROFILES_PARQUET)
-    )
-    company_profiles_json = Path(
-        context.op_config.get("company_profiles_json") or str(DEFAULT_COMPANY_PROFILES_JSON)
-    )
-    batch_size = int(context.op_config.get("batch_size", 1000))
-
-    # Read company profiles
-    expected_cols = (
-        "company_uei",
-        "dominant_cet",
-        "specialization_score",
-        "award_count",
-        "total_funding",
-    )
-    profiles = _read_parquet_or_ndjson(
-        company_profiles_parquet, company_profiles_json, expected_columns=expected_cols
-    )
-    context.log.info(f"Loaded company profiles for Neo4j: {len(profiles)}")
-
-    try:
-        loader = CETLoader(client, CETLoaderConfig(batch_size=batch_size))
-        metrics = loader.load_company_cet_enrichment(profiles)
-        result = {
-            "status": "success",
-            "companies": len(profiles),
-            "metrics": _serialize_metrics(metrics),
-        }
-
-        # Persist a small run summary
-        out_path = DEFAULT_OUTPUT_DIR / "neo4j_company_cet_enrichment.checks.json"
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            with out_path.open("w", encoding="utf-8") as fh:
-                json.dump(result, fh, indent=2)
-        except Exception:
-            pass
-
-        try:
-            client.close()
-        except Exception:
-            pass
-        return result
-    except Exception as exc:
-        context.log.exception(f"Company CET enrichment failed: {exc}")
-        try:
-            client.close()
-        except Exception:
-            pass
-        return {"status": "error", "error": str(exc)}
-
-
-@asset(
-    name="loaded_award_cet_relationships",
-    description="Create Award -> CETArea relationships from award classifications.",
-    group_name="neo4j_cet",
-    ins={
-        "enriched_cet_award_classifications": AssetIn(
-            key=["ml", "enriched_cet_award_classifications"]
-        ),
-        "loaded_cet_areas": AssetIn(),
-        "loaded_award_cet_enrichment": AssetIn(),
-    },
-    config_schema={
-        "award_class_parquet": str,
-        "award_class_json": str,
-        "batch_size": int,
-    },
-)
-def loaded_award_cet_relationships(
-    context, enriched_cet_award_classifications, loaded_cet_areas, loaded_award_cet_enrichment
-) -> dict[str, Any]:
-    """Create Award -> CETArea relationships."""
-    if CETLoader is None or CETLoaderConfig is None:
-        context.log.warning("CETLoader unavailable; skipping Award CET relationships")
-        return {"status": "skipped", "reason": "loader_unavailable"}
-
-    client = _get_neo4j_client()
-    if client is None:
-        return {"status": "skipped", "reason": "neo4j_unavailable"}
-
-    # Config
-    award_class_parquet = Path(
-        context.op_config.get("award_class_parquet") or str(DEFAULT_AWARD_CLASS_PARQUET)
-    )
-    award_class_json = Path(
-        context.op_config.get("award_class_json") or str(DEFAULT_AWARD_CLASS_JSON)
-    )
-    batch_size = int(context.op_config.get("batch_size", 1000))
-
-    # Read award classifications
-    expected_cols = ("award_id", "primary_cet", "supporting_cets", "confidence", "evidence")
-    classifications = _read_parquet_or_ndjson(
-        award_class_parquet, award_class_json, expected_columns=expected_cols
-    )
-    context.log.info(f"Creating Award->CETArea relationships for {len(classifications)} awards")
-
-    try:
-        loader = CETLoader(client, CETLoaderConfig(batch_size=batch_size))
-        metrics = loader.load_award_cet_relationships(classifications)
-        result = {
-            "status": "success",
-            "awards": len(classifications),
-            "metrics": _serialize_metrics(metrics),
-        }
-
-        # Persist a small run summary
-        out_path = DEFAULT_OUTPUT_DIR / "neo4j_award_cet_relationships.checks.json"
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            with out_path.open("w", encoding="utf-8") as fh:
-                json.dump(result, fh, indent=2)
-        except Exception:
-            pass
-
-        try:
-            client.close()
-        except Exception:
-            pass
-        return result
-    except Exception as exc:
-        context.log.exception(f"Award CET relationships failed: {exc}")
-        try:
-            client.close()
-        except Exception:
-            pass
-        return {"status": "error", "error": str(exc)}
-
-
-@asset(
-    name="loaded_company_cet_relationships",
-    description="Create Company -> CETArea relationships from company CET profiles or enrichment.",
-    group_name="neo4j_cet",
-    ins={
-        "transformed_cet_company_profiles": AssetIn(key=["ml", "transformed_cet_company_profiles"]),
-        "loaded_cet_areas": AssetIn(),
-        "loaded_company_cet_enrichment": AssetIn(),
-    },
-    config_schema={
-        "company_profiles_parquet": str,
-        "company_profiles_json": str,
-        "batch_size": int,
-    },
-)
-def loaded_company_cet_relationships(
-    context, transformed_cet_company_profiles, loaded_cet_areas, loaded_company_cet_enrichment
-) -> dict[str, Any]:
-    """Create Company -> CETArea relationships."""
-    if CETLoader is None or CETLoaderConfig is None:
-        context.log.warning("CETLoader unavailable; skipping Company CET relationships")
-        return {"status": "skipped", "reason": "loader_unavailable"}
-
-    client = _get_neo4j_client()
-    if client is None:
-        return {"status": "skipped", "reason": "neo4j_unavailable"}
-
-    # Config
-    company_profiles_parquet = Path(
-        context.op_config.get("company_profiles_parquet") or str(DEFAULT_COMPANY_PROFILES_PARQUET)
-    )
-    company_profiles_json = Path(
-        context.op_config.get("company_profiles_json") or str(DEFAULT_COMPANY_PROFILES_JSON)
-    )
-    batch_size = int(context.op_config.get("batch_size", 1000))
-
-    # Read company profiles
-    expected_cols = (
-        "company_uei",
-        "dominant_cet",
-        "specialization_score",
-        "award_count",
-        "total_funding",
-    )
-    profiles = _read_parquet_or_ndjson(
-        company_profiles_parquet, company_profiles_json, expected_columns=expected_cols
-    )
-    context.log.info(f"Creating Company->CETArea relationships for {len(profiles)} companies")
-
-    try:
-        loader = CETLoader(client, CETLoaderConfig(batch_size=batch_size))
-        metrics = loader.load_company_cet_relationships(profiles)
-        result = {
-            "status": "success",
-            "companies": len(profiles),
-            "metrics": _serialize_metrics(metrics),
-        }
-
-        # Persist a small run summary
-        out_path = DEFAULT_OUTPUT_DIR / "neo4j_company_cet_relationships.checks.json"
-        out_path.parent.mkdir(parents=True, exist_ok=True)
-        try:
-            with out_path.open("w", encoding="utf-8") as fh:
-                json.dump(result, fh, indent=2)
-        except Exception:
-            pass
-
-        try:
-            client.close()
-        except Exception:
-            pass
-        return result
-    except Exception as exc:
-        context.log.exception(f"Company CET relationships failed: {exc}")
-        try:
-            client.close()
-        except Exception:
-            pass
-        return {"status": "error", "error": str(exc)}
-
-
-# ============================================================================
-# Asset Aliases for Backward Compatibility
-# ============================================================================
-
-# Aliases for assets expected by __init__.py and other modules
-neo4j_cetarea_nodes = loaded_cet_areas
-neo4j_award_cet_enrichment = loaded_award_cet_enrichment
-neo4j_company_cet_enrichment = loaded_company_cet_enrichment
-neo4j_award_cet_relationships = loaded_award_cet_relationships
-neo4j_company_cet_relationships = loaded_company_cet_relationships
+__all__ = [
+    # Taxonomy
+    "raw_cet_taxonomy",
+    "cet_taxonomy",
+    "cet_taxonomy_completeness_check",
+    "taxonomy_to_dataframe",
+    # Classifications
+    "enriched_cet_award_classifications",
+    "enriched_cet_patent_classifications",
+    "cet_award_classifications_quality_check",
+    # Training
+    "train_cet_patent_classifier",
+    "cet_award_training_dataset",
+    # Analytics
+    "transformed_cet_analytics",
+    "transformed_cet_analytics_aggregates",
+    # Validation
+    "raw_cet_human_sampling",
+    "validated_cet_iaa_report",
+    "validated_cet_drift_detection",
+    # Company
+    "transformed_cet_company_profiles",
+    "cet_company_profiles_check",
+    # Loading
+    "loaded_cet_areas",
+    "loaded_award_cet_enrichment",
+    "loaded_company_cet_enrichment",
+    "loaded_award_cet_relationships",
+    "loaded_company_cet_relationships",
+    # Backward compatibility aliases
+    "neo4j_cetarea_nodes",
+    "neo4j_award_cet_enrichment",
+    "neo4j_company_cet_enrichment",
+    "neo4j_award_cet_relationships",
+    "neo4j_company_cet_relationships",
+    # Utility functions
+    "save_dataframe_parquet",
+]
